### 2023-11-20 #6072 @editorialbot
[PRE REVIEW]: Tidytacos: An R package for analyses on taxonomic composition of microbial communities
**Submitting author:** @wsmets (<a href="http://orcid.org/0000-0001-5611-6094">Wenke Smets</a>)
**Repository:** https://github.com/LebeerLab/tidytacos
**Branch with paper.md** (empty if default branch): master
**Version:** v0.3.0
