### 2023-11-20 #6072 @kthyng
Hi @wsmets and thanks for your submission. Looking at your note and the package (forked from another repo and having progressed quite a bit from there but still tied to the original repo, and language in the readme "For most functions it is still very basic; this will be improved in the future") it sounds like this package isn't ready to be reviewed. When we ask for reviewer time it is expected that the package is robust and presented its best polished form, everything is ready for outside eyes. Do you need to take more time to wrap up some work before we would send this for review?

We also require a valid open source license.
