### 2023-11-23 cbf6999
Improve documentation
Improve the function documentation for functions in
adders_abundances.R and adders_samples.R.
