+++ R/adders_abundances.R
@@ -1,12 +1,10 @@
-#' Add relative abundance to abundance table
+#' Add relative abundances to count table
 #'
-#' \code{add_rel_abundance} adds relative abundance to the abundance table of a
-#' tidytacos object.
+#' \code{add_rel_abundance} calculates relative abundances of taxa in samples
+#' and adds them to the count table of a tidytacos object under the column name
+#' "rel_abundance".
 #'
-#' This function adds the relative abundance per sample to the abundance table
-#' of a tidytacos object under the variable name "rel_abundance".
-#'
-#' @param ta tidytacos object.
+#' @param ta A tidytacos object.
 #'
 #' @examples
 #' # Initiate count matrix
@@ -23,8 +21,7 @@
 #' )
 #'
 #' # Add relative abundance
-#' data <- data %>%
-#'   add_rel_abundance()
+#' data <- data %>% add_rel_abundance()
 #'
 #' @export
 add_rel_abundance <- function(ta) {
@@ -38,18 +35,17 @@ add_rel_abundance <- function(ta) {
   ta
 }
 
-#' Add absolute abundance to counts table
+#' Add absolute abundances to count table
 #'
-#' \code{add_absolute_abundance} adds absolute abundance to the counts table of a
-#' tidytacos object.
+#' \code{add_absolute_abundance} calculates absolute abundances of taxa in
+#' samples given a taxon that was spiked into all of the samples during library
+#' prep. The function then adds these absolute abundances to the count table of
+#' the tidytacos object under the column name "absolute_abundance".
 #'
-#' This function adds the absolute abundance per taxon and per sample to the counts table
-#' of a tidytacos object under the variable name "absolute_abundance".
-#'
-#' @param ta tidytacos object.
+#' @param ta A tidytacos object.
 #' @param spike_taxon The taxon id of the spike.
-#' @param spike_added The column name of the samples table which indicates how much spike was added per sample, e.g. 16S rRNA gene copy numbers added to the DNA extraction tube.
-#'
+#' @param spike_added The column name in the sample table which indicates how
+#'   much spike was added per sample.
 #'
 #' @examples
 #' # Initiate count matrix
@@ -80,34 +76,41 @@ add_absolute_abundance <- function(ta, spike_taxon, spike_added = spike_added) {
       "that defines the quantity of spike added to the sample."
     ))
   }
-  
-  # if total_counts, relative abundances, and total absolute abundances are not present: add temporarily
+
+  # if total_counts, relative abundances, and total absolute abundances are not
+  # present: add temporarily
   total_counts_tmp <- !"total_counts" %in% names(ta$samples)
   if (total_counts_tmp) ta <- add_total_count(ta)
   rel_abundance_tmp <- !"rel_abundance" %in% names(ta$counts)
   if (rel_abundance_tmp) ta <- add_rel_abundance(ta)
-  total_absolute_abundance_tmp <- !"total_absolute_abundance" %in% names(ta$samples)
-  if (total_absolute_abundance_tmp) ta <- add_total_absolute_abundance(ta, spike_taxon)
-  
+  total_absolute_abundance_tmp <-
+    !"total_absolute_abundance" %in% names(ta$samples)
+  if (total_absolute_abundance_tmp) ta <-
+    add_total_absolute_abundance(ta, spike_taxon)
+
   # make counts table with total abundances
   tot_abs_abundance <- ta$samples %>%
     select(sample_id, total_absolute_abundance)
-  
+
   # calculate total for each taxon in each sample
   ta$counts <- ta$counts %>%
-    left_join(tot_abs_abundance, by='sample_id') %>%
[…]
