+++ R/adders_samples.R
@@ -1,18 +1,16 @@
-#' Add metadata table to the tidytacos object
+#' Add metadata to the tidytacos object
 #'
-#' \code{add_metadata} adds a tibble to the tidytacos object.
+#' \code{add_metadata} adds sample or taxon metadata to the sample or taxon
+#' table, respectively, of a tidytacos object.
 #'
-#' This function adds a tibble containing metadata for each sample or taxon to the
-#' tidytacos object. It is used after initiating a tidytacos object
-#' using a numerical abundance matrix and the function
-#' \code{\link{create_tidytacos}}.
-#'
-#' @param ta tidytacos object.
-#' @param metadata A tibble containing sample data for each sample or taxon. Samples/taxa
-#'   should be rows, while metadata should be columns. At least one column
-#'   name needs to be shared with the sample or taxa tibble of ta. The default shared
-#'   column name is 'sample' for samples and 'taxon' for taxa.
+#' @param ta A tidytacos object.
+#' @param metadata A tibble containing data for each sample or taxon.
+#'   Samples/taxa should be rows, while metadata variables should be columns. At
+#'   least one column name needs to be shared with the sample or taxa table of
+#'   the tidytacos object. The default shared column name is 'sample' for
+#'   samples and 'taxon' for taxa.
 #' @param table_type The type of table to add, either 'sample' or 'taxa'.
+
 #' @examples
 #' # Initiate counts matrix
 #' x <- matrix(
@@ -37,7 +35,7 @@
 #' add_metadata(sample_tibble)
 #'
 #' @export
-add_metadata <- function(ta, metadata_tibble, table_type="sample") {
+add_metadata <- function(ta, metadata_tibble, table_type = "sample") {
 
   if (table_type == "sample") {
     purrr::modify_at(ta, "samples", left_join, metadata_tibble)
@@ -49,15 +47,12 @@ add_metadata <- function(ta, metadata_tibble, table_type="sample") {
 
 }
 
-#' Add total reads per sample
+#' Add total read count per sample
 #'
-#' \code{add_total_count} adds the total reads per sample to the samples tibble
-#' of a tidytacos object.
+#' \code{add_total_count} adds the total read count per sample to the sample
+#' tibble of a tidytacos object under the variable name total_count.
 #'
-#' This function adds the total reads per sample to the samples tibble of a
-#' tidytacos object under the variable name total_count.
-#'
-#' @param ta tidytacos object.
+#' @param ta A tidytacos object.
 #'
 #' @examples
 #' # Initiate counts matrix
@@ -97,18 +92,16 @@ add_total_count <- function(ta) {
 
 }
 
-
 #' Add alpha diversity measures
 #'
-#' \code{add_alpha} adds two alpha diversity measures to the
-#' samples tibble of a tidytacos object.
+#' \code{add_alpha} adds two alpha diversity measures to the sample table of a
+#' tidytacos object.
 #'
 #' This function adds two alpha diversity measures (observed and inverse
-#' Simpson) to the samples tibble of a tidytacos object under the variable
-#' names observed and inverse_simpson, respectively. This function will also
-#' add relative abundances if not present using \code{\link{add_rel_abundance}}.
+#' Simpson) to the sample table of a tidytacos object under the variable
+#' names "observed" and "inverse_simpson", respectively.
 #'
-#' @param ta tidytacos object.
+#' @param ta A tidytacos object.
 #'
 #' @examples
 #' # Initiate counts matrix
@@ -127,6 +120,7 @@ add_total_count <- function(ta) {
 #' # Add total abundance
 #' data <- data %>%
 #'  add_alpha()
+#'
 #' @export
[…]
