+++ man/add_spike_ratio.Rd
@@ -12,15 +12,12 @@ add_spike_ratio(ta, spike_taxon)
 \item{spike_taxon}{The taxon_id of the spike.}
 }
 \description{
-\code{add_spike_ratio} adds a new variable showing the ratio total counts
-to spike counts to the samples tibble of a tidytacos object.
+\code{add_spike_ratio} calculates the ratio of non-spike to spike reads for
+each sample and adds this to the sample table under the name "spike_ratio".
 }
 \details{
-This function calculates the spike ratio defined as the total sample
-counts to the spike counts and adds this as a new variable
-"spike_ratio" to the samples tibble of a tidytacos object. This function
-is useful if a DNA spike was added prior to sequencing and is based on the
-method described by
+This function is useful if a DNA spike was added prior to sequencing and is
+based on the method described by
 \href{https://doi.org/10.1016/j.soilbio.2016.02.003}{Smets et al., 2016}.
 
 Credits to Wenke Smets for the idea of spiking samples prior to 16S
