+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2023-11-13T13:08Z
+last_built: 2023-11-23T14:39Z
 
