+++ .github/workflows/R-CMD-check.yaml
@@ -26,8 +26,10 @@ jobs:
           - {os: macos-latest,   r: 'release'}
 
           - {os: windows-latest, r: 'release'}
-          # Use 3.6 to trigger usage of RTools35
-          - {os: windows-latest, r: '3.6'}
+          # Use 3.6 to trigger usage of RTools35;
+          # latest Matrix package expects R>=4, this is not an issue for most functions
+          # but will fail in the sparcc networking
+          #- {os: windows-latest, r: '3.6'}
           # use 4.1 to check with rtools40's older compiler
           - {os: windows-latest, r: '4.1'}
 
