### 2023-11-24 28d57a9
Add elipsis to adonis
