### 2023-11-27 a1fff0d
Improved function descriptions for the adders.
