### 2023-11-27 cc84106
Function documentation updates to man
