### 2023-11-28 8ad017a
Function description improved in codifab, converters, getters, handlers.
