+++ R/handlers.R
@@ -3,6 +3,9 @@
 #' This function performs rarefying. Make sure that all samples contain at least
 #' the minimum number of reads; otherwise, an error might be thrown.
 #'
+#' @param n Subsample size for rarefying the community.
+#' @param replace Whether to replace the read after it has been selected for the subsample so it can be sampled again? Default is FALSE.
+#'
 #' @export
 rarefy <- function(ta, n, replace = F) {
   ta$counts <-
@@ -22,12 +25,12 @@ rarefy <- function(ta, n, replace = F) {
     process_count_selection()
 }
 
-# Change sample IDs to a given expression
-#
-# @param ta A tidytacos object.
-# @param sample_id_new An expression that evaluates to a unique sample
-#   identifier.
-#
+#' Change sample IDs to a given expression
+#'
+#' @param ta A tidytacos object.
+#' @param sample_id_new An expression that evaluates to a unique sample
+#'   identifier.
+#'
 change_id_samples <- function(ta, sample_id_new) {
   sample_id_new <- rlang::enexpr(sample_id_new)
 
@@ -54,12 +57,12 @@ change_id_samples <- function(ta, sample_id_new) {
   ta
 }
 
-# Change taxon IDs to a given expression
-#
-# @param ta A tidytacos object.
-# @param taxon_id_new An expression that evaluates to a unique taxon
-#   identifier.
-#
+#' Change taxon IDs to a given expression
+#'
+#' @param ta A tidytacos object.
+#' @param taxon_id_new An expression that evaluates to a unique taxon
+#'   identifier.
+#'
 change_id_taxa <- function(ta, taxon_id_new) {
   taxon_id_new <- rlang::enexpr(taxon_id_new)
 
@@ -84,6 +87,8 @@ change_id_taxa <- function(ta, taxon_id_new) {
 }
 
 #' Aggregate samples with identical values for all metadata
+#'
+#' \code{aggregate_samples} merges sample content of samples which have identical values for all columns in the sample table (except sample_id).
 #'
 #' @param ta A tidytacos object.
 #'
@@ -127,8 +132,8 @@ aggregate_samples <- function(ta) {
 #' * If not, delete all taxon variables except taxon_id and the ranks you are
 #' still interested in prior to calling this function.
 #'
-#' @param ta a tidytacos object
-#' @param rank an optional rank to aggregate on
+#' @param ta A tidytacos object.
+#' @param rank An optional rank to aggregate on.
 #' @export
 aggregate_taxa <- function(ta, rank = NULL) {
   if (!is.null(rank)) {
@@ -198,11 +203,12 @@ aggregate_taxa <- function(ta, rank = NULL) {
 
 #' Trim all sequences
 #'
-#' This function assumes that the sequence variable in the taxon table is called
+#' \code{trim_asvs} trims sequence ends of the sequence supplied in the taxa table. This function assumes that the sequence variable in the taxon table is called
 #' "sequence".
-#' @param ta a tidytacos object
-#' @param start index of where to start trimming
-#' @param end index of where to stop trimming
+#'
+#' @param ta A tidytacos object.
+#' @param start Index of where to start trimming.
+#' @param end Index of where to stop trimming.
 #'
 #' @export
 trim_asvs <- function(ta, start, end) {
@@ -221,7 +227,8 @@ trim_asvs <- function(ta, start, end) {
 }
 
 #' Retain or remove a set of sample variables
-#' @param ta a tidytacos object
+#' 
+#' @param ta A tidytacos object.
 #' @export
 select_samples <- function(ta, ...) {
[…]
