### 2023-11-28 1f85527
Allow string isntead of symbol for taxonlis_per_condition
