### 2023-11-28 104c944
Convert to symbol
