### 2023-11-28 82df0b3
Fix bug in class checking
