### 2023-11-29 d0413f1
Remove empty samples if present
