### 2023-11-29 7e58f0b
fixed typo in function description of everything()
