### 2023-12-05 #6072 @wsmets
Well, with addressing points I meant that we added an open source licence, removed the fork to the original repo, we improved function descriptions, and removed the comment from readme relating to function descriptions. We understand your concern, but are confident our package is mature enough for release into the world and therefore also for review.
