### 2023-12-05 e83a2c7
Make sure umap is installed for test
