+++ tests/testthat/test-plotters.R
@@ -46,6 +46,7 @@ test_that("Tacoplot_ord_ly works", {
 
 test_that("Tacoplot_ord_ly works with umap and 3 dims", {
     skip_if_not_installed("plotly")
+    skip_if_not_installed("umap")
     expect_no_error(
         expect_warning(
             urt %>% tacoplot_ord_ly(x=location, ord="umap", dims=3)
