### 2023-12-05 c2de355
Add Mantel test, anosim and interactive option for venndiagram (#17)
* Add ANOSIM test statistic to ord plots

* Check for empty samples prior to analysis

* Copy sequence to taxon column

* Update tests to accomodate new taxon column in urt

* Add files via upload

* Delete joss directory

deleting joss directory from wrong branch

* Add mantel test

* Add code comments

* Improve documentation

Improve the function documentation for functions in
adders_abundances.R and adders_samples.R.

* Improved function descriptions for the adders.

* Function documentation updates to man

* Warn if and 
[…]
