+++ R/plotters.R
@@ -30,16 +30,23 @@ prepare_for_bp <- function(ta, n = 12, extended = TRUE) {
 }
 
 #' Return a bar plot of the samples
+#' 
+#' Plots a stacked bar plot of the samples in the tidytacos object to inspect the taxonomic profile.
+#' 
+#' @param ta A tidytacos object.
+#' @param x A string, representing the column name used to label the x-axis
+#' @param n An integer, representing the amount of colors used to depict
+#' @param geom_bar A boolean, whether or not to add geom_bar to the plot. Default is TRUE.
 #'
 #' @export
 tacoplot_stack <- function(ta, n = 12, x = sample_clustered, geom_bar = T) {
   # convert promise to formula
-  x <- enquo(x)
+  x <- rlang::enquo(x)
 
-  warning_message_label = paste0("Label \'", quo_name(x),"\' not found in the samples table.")
+  warning_message_label = paste0("Label \'", rlang::quo_name(x),"\' not found in the samples table.")
   warning_message_aggregate = "Sample labels not unique, samples are aggregated."
-  if (quo_name(x) != "sample_clustered" &&
-    !is.element(quo_name(x), names(ta$samples))
+  if (rlang::quo_name(x) != "sample_clustered" &&
+    !is.element(rlang::quo_name(x), names(ta$samples))
   ) {
     # Warning, so tidy functions can be performed on the label
     warning(warning_message_label)
@@ -83,11 +90,12 @@ tacoplot_stack <- function(ta, n = 12, x = sample_clustered, geom_bar = T) {
 
 #' Return an interactive bar plot of the samples
 #'
+#' Plots an interactive stacked bar plot of the samples in the tidytacos object to inspect the taxonomic profile.
+#' 
 #' @param ta A tidytacos object.
 #' @param n An integer, representing the amount of colors used to depict
 #'   different taxa.
-#' @param x A string, representing the column name used to label and cluster the
-#'   samples on.
+#' @param x A string, representing the column name used to label the x-axis
 #'
 #' @export
 tacoplot_stack_ly <- function(ta, n = 12, x = sample_clustered) {
@@ -122,7 +130,10 @@ tacoplot_stack_ly <- function(ta, n = 12, x = sample_clustered) {
   plot
 }
 
-#' Return an interactive pcoa plot of the samples
+#' Return an interactive ordination plot of the samples
+#' 
+#' Creates an interactive ordination plot of the beta diversity of the samples in the tidytacos object.
+#' This can be used to gauge the similarity between samples.
 #'
 #' @param ta A tidytacos object.
 #' @param x A string, representing the column name used to color the sample
@@ -131,12 +142,14 @@ tacoplot_stack_ly <- function(ta, n = 12, x = sample_clustered) {
 #' @param ord the ordination technique to use. Choice from pcoa, tsne and umap.
 #' @param distance the distance algorithm to use, see \code{\link[vegan]{vegdist}}.
 #' @param dims the amount of dimensions to plot, 2 or 3.
+#' @param stat.method the statistic to print on the figure, choice from mantel and anosim.
 #' @param palette A vector of colors, used as the palette for coloring sample
 #' @param title a string to display as title of the plot.
 #'   groups.
 #'
 #' @export
-tacoplot_ord_ly <- function(ta, x=NULL, samplenames = sample_id, ord="pcoa", dims=2, distance="bray", palette = NULL, title = NULL, ...) {
[…]
