+++ _pkgdown.yml
@@ -1,3 +1,4 @@
 template:
   bootstrap: 5
   bootswatch: flatly
+  url: https://lebeerlab.github.io/tidytacos/
