+++ pkgdown.yml
@@ -2,5 +2,8 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2023-12-05T14:28Z
+last_built: 2023-12-05T14:40Z
+urls:
+  reference: https://lebeerlab.github.io/tidytacos/reference
+  article: https://lebeerlab.github.io/tidytacos/articles
 
