+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2023-12-05T14:40Z
+last_built: 2023-12-05T14:49Z
 urls:
   reference: https://lebeerlab.github.io/tidytacos/reference
   article: https://lebeerlab.github.io/tidytacos/articles
