### 2023-12-12 ff4fb04
Add function to write sequences to fasta
