+++ NAMESPACE
@@ -80,6 +80,8 @@ export(tacosum)
 export(taxa)
 export(taxonlist_per_condition)
 export(test_taco)
+export(to_biom)
+export(to_fasta)
 export(trim_asvs)
 export(write_tidytacos)
 import(dplyr)
