### 2024-01-09 6383b9e
Rename column output of add_prevalence
