+++ tests/testthat/test-plotters.R
@@ -56,7 +56,7 @@ test_that("Tacoplot_ord_ly works with umap and 3 dims", {
 })
 
 test_that("Can create venndiagram", {
-    skip_if_not_installed("ggVenDiagram")
+    skip_if_not_installed("ggVennDiagram")
     venn <- urt %>% tacoplot_venn(location)
     vdiffr::expect_doppelganger("Venndiagram", venn)
 })
