+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2023-12-05T14:49Z
+last_built: 2024-01-11T11:59Z
 urls:
   reference: https://lebeerlab.github.io/tidytacos/reference
   article: https://lebeerlab.github.io/tidytacos/articles
