### 2024-01-19 #6072 @mikemahoney218
Sorry @kthyng and @SWittouck , I think this is a bit too far afield for me -- I don't have any experience with microbes and last touched a microscope during my second year of undergrad. I need to turn this one down unfortunately.
