### 2024-01-24 52a9dfe
Add stat to plot and style it a little
