### 2024-02-01 #6072 @david-barnett
> @david-barnett and @kelly-sovacool: Would you be available, and interested, in reviewing this JOSS submission on microbial community data?

I'm in 👍
