### 2024-02-01 #6313 @editorialbot
[REVIEW]: Tidytacos: An R package for analyses on taxonomic composition of microbial communities
**Submitting author:** @wsmets (<a href="http://orcid.org/0000-0001-5611-6094">Wenke Smets</a>)
**Repository:** https://github.com/LebeerLab/tidytacos
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.6
