### 2024-02-01 #6313 @diazrenata
@kelly-sovacool, @david-barnett - Thank you for agreeing to review this submission! We'll continue the conversation on the review here.

In case it is helpful, here is the JOSS reviewers' guide: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Thank you both!
