### 2024-02-06 c8b59ba
Can supply vector of alpha metrics
