### 2024-02-18 #6313 @david-barnett
Hello all, 👋

I started to work through the reviewer checklist today. I have not completed my review of tidytacos but I have some questions and comments already.

Points 1 and 2 seem to me like practical issues worth addressing before proceeding with further review. Points 3 and 4 refer to issues with the manuscript and package documentation, which if addressed would make it easier to continue with a comprehensive and fair review of the package.

## 1. Checking contribution of submitting author

From the JOSS reviewer guidelines: <https://joss.readthedocs.io/en/latest/review_criteria.html#authorship>

> As part of the review process, you are asked to check whether the submitting author has made a ‘substantial contribution’ to the submitted software (as determined by the commit history)

Wenke Smets is the submitting author, but I did not find evidence of code contributions from this author.

From the commit history on github, I initially found only two commits from @wsmets, adding the manuscript for JOSS submission. <https://github.com/LebeerLab/tidytacos/commits/dev?author=wsmets>

I later found a couple more commits authored by "wsmets" but not linked to a github account (and so not easily searchable). What I found seemed to be small updates only to the documentation of several functions. <https://github.com/LebeerLab/tidytacos/pull/17/commits/8ad017a838e93576823113e30b9a8fabf51e6a0a>

@wsmets , is this a fair assessment of your contributions to tidytacos? If not, could you clarify what I missed?

## 2. Version for review? dev branch is 61 commits ahead of master

It looks like the dev branch is quite active, and seems to include several features and documentation updates? It would be good to get clarity on the short term plans for this development. e.g. should this all be merged to master and tagged as a new version before further review?

## 3. Reference to similar work

From JOSS [reviewer checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) :

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

From JOSS [review criteria page](https://joss.readthedocs.io/en/latest/review_criteria.html#an-important-note-about-novel-software-and-citations-of-relevant-work):

> Submissions that implement solutions already solved in other software packages are accepted into JOSS provided that they meet the criteria listed above and cite prior similar work. Reviewers should point out relevant published work which is not yet cited.

Currently the manuscript does not cite or mention similar work. tidytacos provides a generalist framework for handling data about the taxonomic composition of microbiomes, and facilitating the application of various relevant analyses and visualisations. Here are some R packages that address similar aims:

-   phyloseq and packages using phyloseq objects: microbiome, microViz, phylosmith
-   treeSummarizedExperiment and packages using that object class: microbiotaProcess, mia & miaViz
-   Others: microeco, ampvis2, microbiomeanalystR

It would be great to see a few words added to the manuscript acknowledging related work. I think there is no need or space for an extensive comparison, but given that there are now so many related packages with overlapping functionality, it would be good to see some justification for the usefulness of tidytacos included in the paper. The philosophy section of the wiki claims that the simplicity of tidytacos' design provides some advantages over packages based around a more complex object structure e.g. phyloseq. I think this is a reasonable line of argument to expand upon in the paper.

From JOSS [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort):

> Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).

## 4. Package documentation and other user guidance

I installed tidytacos from the master branch and ran all the wiki tutorial material locally.

<https://github.com/LebeerLab/tidytacos/wiki/Tutorials>

This tutorial started by demonstrating: making stacked bar charts; plotting observed ASV richness; making a 3D PCoA using Bray-Curtis dissimilarities computed on genus-level aggregated features. That worked fine, and it was mostly clear enough how I could make some simple changes to modify these procedures.

However, I have some suggestions for the documentation overall:

1.  There appears to be a substantial amount of tidytacos functionality not covered by tutorial material. For example, there are several functions related to network analyses in the [network.R file](https://github.com/LebeerLab/tidytacos/blob/HEAD/R/network.R). These have minimal docs, without examples of use. Network analysis functionality is not mentioned in the manuscript either. I think there should be a more comprehensive structured overview available of the package's functionality available, to help guide new users. As well as adding tutorial material or other examples of use for more of the functionality, the authors could also consider structuring the reference page by theme, instead of listing all \~80 functions alphabetically.

2.  In the Differential Abundance Analysis [section of the tutorial wiki](https://github.com/LebeerLab/tidytacos/wiki/Tutorials#differential-abundance-analysis) - the add_codifab and tacoplot_codifab functions are introduced. I do not recall seeing this approach for statistical analysis and visualisation before. It would be great to get a reference if this is an established method? Regardless, the method could be better described and justified in the documentation, as the details are currently unclear without looking at the underlying code. None of the codifab.R functionality is mentioned in the manuscript either.
