### 2024-02-26 43b6ba8
Fix clustering bug when sample_name column is present
