### 2024-03-08 #6313 @wsmets
@david-barnett, @kelly-sovacool, and @diazrenata, thank you so much for taking the time to look into our tool and considering its publication! As a response to the first points raised by @david-barnett:

1. As for the limited contribution of the submitting author, I am actually new to GitHub and never tracked my changes to the package before. Besides the improvements to function descriptions, my greatest code contribution consists of four functions regarding the transformation of data to absolute abundances in case a spike was added. This is, to the best of our knowledge, an entirely new feature to the tools of amplicon sequencing analyses, but these four functions compared to the whole package remain a limited one. My other contributions consist more of giving user-based input, joining in meetings of package conceptualization, and supporting the organization and the writing of this publication, hence my role as submitting author. However, if these contributions are considered insufficient, we don’t mind changing the submitting author. @diazrenata, in that case, please let us know how we should proceed.

2. We are actively maintaining and expanding the package as user-based feedback comes in. We think the package can be reviewed as such, without considering the dev branch.

3. We think it is a valuable addition to mention existing work as you propose, and we are currently updating the manuscript to include the related packages.

4. - More tutorials are indeed a valid point of improvement. We are making additional tutorials and will update you here once they are available.
    - Following your other suggestion, a list of functions sorted thematically was added to the github page here: https://lebeerlab.github.io/tidytacos/reference/index.html#file-handling
    - We are currently also working on a clarification of the principles of codifab in the manuscript.

We want to thank you for your patience while we are implementing your feedback in the package and the manuscript.
