### 2024-03-08 708aca9
Prevent clustering of single sample
