### 2024-03-27 61ace81
Add optional order_by argurment to tacoplot_stack
