+++ man/tacoplot_stack_ly.Rd
@@ -4,7 +4,7 @@
 \alias{tacoplot_stack_ly}
 \title{Return an interactive bar plot of the samples}
 \usage{
-tacoplot_stack_ly(ta, n = 12, x = sample_clustered)
+tacoplot_stack_ly(ta, n = 12, x = sample_clustered, ...)
 }
 \arguments{
 \item{ta}{A tidytacos object.}
@@ -13,6 +13,9 @@ tacoplot_stack_ly(ta, n = 12, x = sample_clustered)
 different taxa.}
 
 \item{x}{A string, representing the column name used to label the x-axis}
+
+\item{order_by}{an optional column name to order the samples by. 
+For examples order_by=sample would order the x-axis by the sample names instead of by similar profiles.}
 }
 \description{
 Plots an interactive stacked bar plot of the samples in the tidytacos object to inspect the taxonomic profile.
