### 2024-03-27 b6e138c
Fix bug in tacoplot_stack ellipsis
