### 2024-03-27 523b681
Add wiki articles as vignettes
