+++ vignettes/stacked-barplot.Rmd
@@ -1,0 +1,175 @@
+---
+title: "Stacked Barplots"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{stacked-barplot}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+This vignette explains in more detail how to modify the stacked bar
+plots of tidytacos. We will illustrate those using a dataset with human
+microbiome samples from the upper respiratory tract (URT), taken from
+[this
+paper](https://www.frontiersin.org/articles/10.3389/fmicb.2017.02372/full)
+by De Boeck et al. It contains nose as well as nasopharynx samples. Most
+samples were taken using a swab method, but a minority was taking with
+the aspirate method.
+
+## Setting up
+
+We need only two packages: tidytacos (of course) and the tidyverse set
+of packages.
+
+```{r, setup}
+library(tidyverse)
+library(tidytacos)
+```
+
+## Making a stacked barplot of a subset of samples
+
+We start by selecting the samples we are interesting in: only the ones
+from the nose, taken with the swab method.
+
+```{r}
+urt_ns <- urt %>%
+  filter_samples(location == "N", method == "S")
+```
+
+We can very easily create explorative plots of our samples in the
+following ways:
+
+```{r, stackplot, fig.width=8,fig.height=5}
+tacoplot_stack(urt_ns)
+```
+
+```{r interactive-stackplot, fig.width=8,fig.height=5}
+tacoplot_stack_ly(urt_ns, x=sample)
+```
+
+The `tacoplot_stack` function returns a nice bar plot visualization of
+the most abundant taxa in our samples. The `tacoplot_stack_ly` function
+returns an interactive version of the same bar plot. The order of the
+samples on the x-axis is determined by hierarchical clustering of the
+visualized sample composition. In addition, these functions do some
+things behind the screens:
+
+- Add relative abundances to the counts table.
+- For each taxon, calculate the mean relative abundance across all
+  samples where this taxon occurs.
+- Give all taxa a human-understandable name so that the taxon name is
+  unique. This is just the genus name of the taxon, followed by a number
+  to make it unique. Taxa with a higher mean relative abundance get a
+  smaller number. E.g. “Lactobacillus 1” is the Lactobacillus taxon with
+  the largest mean relative abundance.
+- Make a new variable that is equal to the taxon name, except that only
+  the top-11 taxa retain their name and all others are changed to “Other
+  taxa”. This is for visualization purposes; the human eye can only
+  discriminate clearly between about 12 colors.
+
+All these new variables are created under the hood, but are gone when
+the function execution is finished (there are no so-called
+“side-effects”). Luckily, for each of these variables there also exists
+a function that will create it and keep it! These functions are the
+following (there names are intended to be self-explanatory as much as
+possible):
+
+- `add_rel_abundance`: adds to counts table
+- `add_mean_rel_abundance`: adds to taxa table
+- `add_taxon_name`: adds to taxa table
+- `add_taxon_name_color`: adds to taxa table
+
+There are more of such functions that add an extra variable to one of
+the tables; we will explore them further in other parts of this
+vignette.
+
+## Making a facet wrapped stacked barplot
+
+There are two interesting types of facet wrapped bar plots we can make:
+facets with fixed x-axis categories, and facets without fixed x-axis
+categories.
+
+**Facets with fixed x-axis categories**
+
+The first type is a plot where the variable on the x-axis is not the
[…]
