+++ vignettes/getting-started.Rmd
@@ -1,0 +1,257 @@
+---
+title: "Tidytacos quick start guide"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{getting-started}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+This guide explains how to perform some common microbial sequencing
+count data analysis tasks using tidytacos. We will illustrate those
+using a dataset with human microbiome samples from the upper respiratory
+tract (URT), taken from [this
+paper](https://www.frontiersin.org/articles/10.3389/fmicb.2017.02372/full)
+by De Boeck et al. It contains nose as well as nasopharynx samples. Most
+samples were taken using a swab method, but a minority was taking with
+the aspirate method.
+
+## Philosophy of the package
+
+A tidytacos object is simply a list of three tables:
+
+- **counts**: These are the counts of reads for each taxon
+  (OTU/ASV/phylotype) in each sample. Each row represents such a read
+  count.
+- **samples**: This table contains the sample metadata. Each row
+  represents a sample.
+- **taxa**: This table contains the taxonomy and other metadata for the
+  taxa. Each row represents a taxon.
+
+The package is called tidytacos because each of the tables is tidy: every
+row represents an observation and every column a variable (more on data
+tidying can be found at
+<https://r4ds.hadley.nz/data-tidy.html#sec-tidy-data>). For this quick
+start guide, we assume you are familiar with the tidyverse (especially
+dplyr and ggplot).
+
+The main differences with the popular phyloseq package are:
+
+- The abundance table of tidytacos is in long format, while the
+  otu_table object of phyloseq is in wide (matrix) format. A tidy read
+  count table is more compact for very sparse data (such as microbiome
+  data), and also easier to handle and visualize within the tidyverse
+  framework.
+- All tables in tidytacos are data frames (technically tibbles). In
+  phyloseq, each table has its own special data type (i.e. the classes
+  “otu_table”, “tax_table” and “sample_data”). This makes it sometimes
+  difficult to access the data directly and do simple stuff with it.
+  Also, the otu_table and tax_table classes are based on the matrix
+  type, while the sample_data class is based on the dataframe type. This
+  also makes data manipulation sometimes un-straightforward.
+- In a phyloseq otu table, sometimes the rows are the taxa and sometimes
+  the columns. This can easily lead to errors.
+
+## Setting up
+
+In case you haven’t installed tidytacos yet, it can be installed using
+devtools:
+
+```{r install, eval=FALSE}
+install.packages("devtools")
+devtools::install_github("LebeerLab/tidytacos")
+```
+
+For this guide, we only need to load two packages: tidytacos (of course) and the tidyverse set
+of packages.
+
+```{r setup}
+library(tidyverse)
+library(tidytacos)
+```
+
+The first step will usually be the conversion our data into a usable
+tidytacos format. If, for example, you used DADA2 as explained
+[here](http://rstudio-pubs-static.s3.amazonaws.com/273862_8258982fdfac49a8ba9d6ed651b9b539.html),
+you can hand off the data to tidytacos using `from_dada`. You can also
+convert a phyloseq object to a tidytacos object using the `as_tidytacos`
+function. More options to import and convert your data can be found
+[here](https://github.com/LebeerLab/tidytacos/blob/dev/R/converters.R).
+
+A tidytacos object is read and stored as three sparse tables (counts-, taxa- and samples.csv). 
+To read in existing data from a folder called 'my_data' you would run:
+
[…]
