### 2024-03-27 e3d4190
Add first draft network vignette
