### 2024-03-27 65f2ff2
Change figsizes
