+++ vignettes/network-analysis.Rmd
@@ -1,0 +1,123 @@
+---
+title: "network-analysis"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{network-analysis}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(tidyverse)
+library(tidytacos)
+
+urt.nf <- urt %>% 
+filter_samples(location == "NF") %>%
+add_taxon_name() 
+
+```
+
+This is a brief guide on how to perform network analysis on microbiome
+count data using a similar strategy as performed in the [Isala study](https://www.nature.com/articles/s41564-023-01500-0) using tidytacos. 
+
+The method clusters taxa based on their co-occurrence patterns across the samples. 
+
+As per usual the test dataset used is a human microbiome samples from the upper respiratory
+tract (URT), taken from [this
+paper](https://www.frontiersin.org/articles/10.3389/fmicb.2017.02372/full)
+by De Boeck et al. It contains nose as well as nasopharynx samples. Most
+samples were taken using a swab method, but a minority was taking with
+the aspirate method.
+
+## Step 1: Creating a network
+
+The first step in the analysis is creating a sparcc network from our taxonomic counts.
+For this the [SpiecEasi package](https://github.com/zdk123/SpiecEasi) is used.
+
+```{r, install-spieceasi, eval=FALSE}
+devtools::install_github('zdk123/SpiecEasi')
+```
+
+```{r, sparcc-network }
+urt.net <- urt.nf %>%
+  network(
+    rarefact = 0.05, # rarefaction level
+    iter=20, # number of iterations of the outer loop of sparcc 
+    inner_iter=10, # number of iterations of the inner loop of sparcc
+    th=0.1, # correlations beneath this threshold are set to 0
+    taxon_name=taxon_name)
+
+# The network object is a list with two elements:
+urt.net$Cor %>% head(2)# Correlation matrix
+# urt.net$Cov (covariance matrix)
+# urt.net$names (names of the taxa on x and y-axes)
+```
+
+# Step 2: Filtering the network matrix
+
+Next up we filter the output correlation matrix to remove weak correlations.
+For this we use the `filter_network` function, which has the `Matrix` package as a dependency.
+We also install the `igraph` package, which can be used for the visualization of the network.
+```{r, eval=FALSE}
+install.packages("Matrix")
+install.packages("igraph")
+```
+
+We remove weak correlations (here set to 0.1) take the absolute value of the correlation matrix and remove the diagonal.
+This is all handled with the `filter_network` function, but we show the code here for clarity and to plot the network graph before clustering.
+```{r, network-filter, fig.width=7, fig.height=7}
+library(Matrix)
+library(igraph)
+
+urt.net.filt <- urt.net %>% filter_network(threshold = 0.1)
+# this is what the function does:
+se.sparcc.graph <- abs(urt.net$Cor) >= 0.1 # the treshold, absolutes as analysis is undirected
+diag(se.sparcc.graph) <- 0 # remove diagonal correlations (= always 1)
+sparcc.graph <- Matrix::Matrix(se.sparcc.graph, sparse = TRUE)
+rownames(sparcc.graph) <- urt.net$names
+graph <- igraph::graph_from_adjacency_matrix(sparcc.graph, mode = "undirected")
[…]
