### 2024-03-28 3c68bf3
Add dependencies needed for vignettes
