+++ .github/workflows/pkgdown.yaml
@@ -29,23 +29,22 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: |
-          any::pkgdown 
-          local::.
-          any::tidyverse
-          any::plotly
-          any::vdiffr
-          any::compositions
-          any::MCL
-          any::Matrix
-          any::umap
-          any::Rtsne
-          any::fastDummies
-          any::ggVennDiagram
-          bioc::dada2
-          bioc::phyloseq
-          github::zdk123/SpiecEasi
-
-        needs: website
+            any::pkgdown
+            local::.
+            any::tidyverse
+            any::plotly
+            any::vdiffr
+            any::compositions
+            any::MCL
+            any::Matrix
+            any::umap
+            any::Rtsne
+            any::fastDummies
+            any::ggVennDiagram
+            bioc::dada2
+            bioc::phyloseq
+            github::zdk123/SpiecEasi
+          needs: website
 
       - name: Build site
         run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
