### 2024-03-28 1e833a2
Add packages used in vignettes to suggests
