+++ DESCRIPTION
@@ -23,7 +23,8 @@ Imports:
     tibble,
     tidyr,
     vegan
-Suggests: 
+Suggests:
+    tidyverse, 
     covr,
     compositions,
     vdiffr,
@@ -32,6 +33,9 @@ Suggests:
     knitr,
     phyloseq,
     plotly,
+    MCL,
+    Matrix,
+    zdk123/SpiecEasi,
     rmarkdown,
     roxygen2,
     testthat (>= 3.1.9)
