### 2024-03-28 81ef83d
Add igraph to suggest
