+++ tests/testthat/test-adders_taxa.R
@@ -61,15 +61,15 @@ test_that("Prevalence can not be higher than amount of samples", {
     expect_lte(max(ta_occ$taxa$prevalence), dim(urt$samples)[1])
 })
 
-test_that("Prevalence in conditions with fischer test can be run", {
-    ta_occ <- urt %>% add_prevalence(condition="location", fischer_test=TRUE)
+test_that("Prevalence in conditions with fisher test can be run", {
+    ta_occ <- urt %>% add_prevalence(condition="location", fisher_test=TRUE)
     expect_true(all(c("prevalence_in_N","prevalence_in_NF","fisher_p") %in% names(ta_occ$taxa)))
     expect_lte(max(ta_occ$taxa$prevalence_in_NF), dim(urt$samples)[1])
     expect_lte(max(ta_occ$taxa$prevalence_in_N), dim(urt$samples)[1])
 })
 
-test_that("Relative prevalences in conditions with fischer test can be run", {
-    ta_occ <- urt %>% add_prevalence(condition="location", fischer_test=TRUE)
+test_that("Relative prevalences in conditions with fisher test can be run", {
+    ta_occ <- urt %>% add_prevalence(condition="location", fisher_test=TRUE)
     expect_true(all(c("prevalence_in_N","prevalence_in_NF","fisher_p") %in% names(ta_occ$taxa)))
     expect_lte(max(ta_occ$taxa$prevalence_in_NF), dim(urt$samples)[1])
     expect_lte(max(ta_occ$taxa$prevalence_in_N), dim(urt$samples)[1])
