+++ man/cluster_network.Rd
@@ -12,7 +12,7 @@ cluster_network(network, min_n = 3, visualize = F)
 
 \item{min_n}{minimum number of taxa per cluster, smaller clusters are filtered out.}
 
-\item{visualize}{wether to visualize the network clusters or not.}
+\item{visualize}{whether to visualize the network clusters or not.}
 }
 \description{
 Performs Markov Clustering on a sparcc network matrix generated by \code{\link{filter_network}}. 
