### 2024-04-04 6e89b53
Remove self loops
