+++ vignettes/data-processing.Rmd
@@ -2,7 +2,11 @@
 title: "Tutorial data processing"
 author: "Wenke Smets"
 date: "2024-04-03"
-output: html_document
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{data-processing}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
 ---
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE, message = F, warning = F, fig.width = 8, fig.height = 6)
@@ -16,7 +20,7 @@ Tutorial data processing
 This tutorial walks you through some quality checks when first diving into a dataset of taxonomic compositions. Note that this tutorial does not include sequence quality filtering, as we are assuming this was already done by for example DADA2. Here we will focus on filtering the dataset to keep only bacterial reads, inspect blanks, and removing samples with too few reads. How you handle these steps may depend on your dataset, but here's a demonstration to give you an idea of your options.
 
 
-### Seleafing up
+### Setting up
 
 We need only two packages: tidytacos (of course) and the tidyverse set of packages. 
 
@@ -81,7 +85,7 @@ ggplot(leaf$samples, aes(x = log10(total_count), y = sample, color = Plant))+
 
 ```
 
-Unfortunately, one sample (boleafom right corner) took up a lot of the reads of this sequencing run, decreasing sampling depth of all other samples. Nevertheless, most samples have more than 1000 reads, which we consider worthwhile. Blanks are on the low end of the spectrum in general, which is according to our expectation, but some blanks have quite some reads which requires further investigation.
+Unfortunately, one sample (bottom right corner) took up a lot of the reads of this sequencing run, decreasing sampling depth of all other samples. Nevertheless, most samples have more than 1000 reads, which we consider worthwhile. Blanks are on the low end of the spectrum in general, which is according to our expectation, but some blanks have quite some reads which requires further investigation.
 
 Let's look at the composition of our blanks. We use the `filter_samples` function to generate a tidytacos object with only the blanks. The `tacoplot_stack` function returns a nice stacked bar plot visualization of the most abundant taxa in the blanks.
 
@@ -96,14 +100,15 @@ tacoplot_stack(leaf1)+
 
 ```
 
-Citrobacter is suspiciously abundant and occurs in mulitple blanks! After checking its sequence, this ASV is the one that we spiked into our samples and many of our blanks to determine absolute bacterial numbers from the relative abundances. Hence, in this case, the high abundance of Citrobacter in the blanks is a good thing. Besides Citrobacter none of the contaminants are consistent nor do the read numbers of the blanks without spike exceed 1000 per sample, which are all good signs for the quality of our data.
+Citrobacter is suspiciously prevalent and occurs in multiple blanks! After checking its sequence, this ASV is the one that we spiked into our samples and many of our blanks to determine absolute bacterial numbers from the relative abundances. Hence, in this case, the high abundance of Citrobacter in the blanks is a good thing. Besides Citrobacter none of the contaminants are consistent nor do the read numbers of the blanks without spike exceed 1000 per sample, which are all good signs for the quality of our data.
 
-We still need to set a minimum read number that a sample needs to have to be included in our analyses. Samples with really low read counts are too likely not being representative of the sampled microbiome by poor amplification, contamination, and/or insufficient sequencing depth. Usually, a minimum read number cut-off is quite an arbitrary decision, hopefully somewhat informed by the nature of your data and results from the blanks. In this case however, we have spiked our samples and several of our blanks and we can determine a threshold that is actually informed by the amount of contaminants in the blanks! 
[…]
