### 2024-04-04 c22dd36
Remove self loop
