+++ pkgdown.yml
@@ -1,8 +1,12 @@
 pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
-articles: {}
-last_built: 2024-03-27T09:33Z
+articles:
+  data-processing: data-processing.html
+  getting-started: getting-started.html
+  network-analysis: network-analysis.html
+  stacked-barplot: stacked-barplot.html
+last_built: 2024-04-04T21:46Z
 urls:
   reference: https://lebeerlab.github.io/tidytacos/reference
   article: https://lebeerlab.github.io/tidytacos/articles
