### 2024-04-09 59c99d5
Add permanova test to possible pcoa annotations
