+++ man/counts_matrix.Rd
@@ -4,7 +4,12 @@
 \alias{counts_matrix}
 \title{Return a counts matrix}
 \usage{
-counts_matrix(ta, sample_name = sample, taxon_name = taxon, value = count)
+counts_matrix(
+  ta,
+  sample_name = sample_id,
+  taxon_name = taxon_id,
+  value = count
+)
 }
 \arguments{
 \item{ta}{A tidytacos object.}
