+++ man/tacoplot_euler.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plotters.R
+\name{tacoplot_euler}
+\alias{tacoplot_euler}
+\title{Return an euler diagram of overlapping taxon_ids between conditions}
+\usage{
+tacoplot_euler(ta, condition, shape = "ellipse", ...)
+}
+\arguments{
+\item{ta}{A tidytacos object.}
+
+\item{condition}{The name of a variable in the samples table that contains a
+categorical value.}
+
+\item{shape}{shape to plot the groups in; choice from circle or ellipse}
+}
+\description{
+Return an euler diagram of overlapping taxon_ids between conditions
+}
