+++ man/tidytacos.Rd
@@ -2,6 +2,7 @@
 % Please edit documentation in R/tidytacos.R
 \docType{package}
 \name{tidytacos}
+\alias{-package}
 \alias{tidytacos}
 \title{tidytacos: Functions to manipulate and visualize amplicon count data.}
 \description{
