+++ DESCRIPTION
@@ -44,7 +44,7 @@ VignetteBuilder:
     knitr
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 Config/testthat/edition: 3
 URL: https://lebeerlab.github.io/tidytacos/
 BugReports: https://github.com/LebeerLab/tidytacos/issues/new?assignees=TheOafidian&labels=bug&projects=&template=bug_report.md&title=%5BBUG%5D+
