### 2024-05-13 #6313 @wsmets
Dear [@david-barnett](https://github.com/david-barnett), [@kelly-sovacool](https://github.com/kelly-sovacool), and [@diazrenata](https://github.com/diazrenata),

Sorry for the delay, but we have further addressed the points raised by [@david-barnett](https://github.com/david-barnett):

- We adapted the text to include some of the most similar packages:
“` tidytacos` offers similar functionalities as packages such as microViz (@barnett2021microviz), mia (@ernst1mia), and phyloseq (@mcmurdie2013phyloseq), but was designed specifically to be very accessible to researchers with limited R and general programming experience. To achieve this, tidytacos was embedded in the tidyverse framework, which provides a style of programming that is very intuitive for new (but also experienced) programmers and is often the starting point for many people learning R. Tidytacos is also very easy to install, since it has only the main tidyverse packages as dependencies for the vast majority of its functionality.  Finally, tidytacos also serves the needs of researchers who are looking to implement some rarely used or new visualizations and methods of analysis.”

-  We have added tutorials for the network.R functions and for data processing, which can be found here: https://lebeerlab.github.io/tidytacos/articles/index.html

- The “codifab” functions used to find taxa that are more or less abundant in one condition versus another, are based on a rather old but solid principle of differences between logratios first proposed by Aitchison. It is for example also used by ANCOM. We have expanded the descriptions of the functions “add_codifab” and “tacoplot_codifab” to improve interpretation of the results and added Aitchisons original paper. We furthermore added to the text of the manuscript:
“and differentially abundant taxa can be determined with a method based on the logratio differences introduced by @aitchison1982statistical”
