+++ man/network.Rd
@@ -2,35 +2,36 @@
 % Please edit documentation in R/network.R
 \name{network}
 \alias{network}
-\title{Perform network inference with SparCC on a tidytacos object, 
+\title{Perform network inference with SparCC on a tidytacos object,
 after dropping rare taxa. See \link[SpiecEasi]{sparcc}.}
 \usage{
 network(
   ta,
-  calculate_p = TRUE,
   min_occurrence = 0.05,
   taxon_name = taxon,
   sample_name = sample,
+  calculate_p = FALSE,
   ...
 )
 }
 \arguments{
 \item{ta}{a tidytacos object}
 
-\item{calculate_p}{whether to calculate p-values or not. 
-This can be time consuming. 
-Iterations can be set with the R parameter and multiple cores through ncpus.}
+\item{min_occurrence}{Percentage of samples the taxon
+needs to be present in for it to be considered in the analysis.}
 
-\item{min_occurrence}{Percentage of samples the taxon needs to be present in for it to be considered in the analysis.}
+\item{calculate_p}{whether to calculate p-values or not.
+This can be time consuming due to the many iterations needed.
+Iterations can be set with the R parameter and multiple cores through ncpus.}
 
 \item{iter}{Number of iterations in the outer loop}
 
 \item{inner_iter}{Number of iterations in the inner loop}
 
-\item{th}{absolute value of correlations below this threshold are 
+\item{th}{absolute value of correlations below this threshold are
 considered zero by the inner SparCC loop.}
 }
 \description{
-Perform network inference with SparCC on a tidytacos object, 
+Perform network inference with SparCC on a tidytacos object,
 after dropping rare taxa. See \link[SpiecEasi]{sparcc}.
 }
