+++ NAMESPACE
@@ -74,6 +74,7 @@ export(tacoplot_codifab)
 export(tacoplot_euler)
 export(tacoplot_ord)
 export(tacoplot_ord_ly)
+export(tacoplot_prevalences)
 export(tacoplot_stack)
 export(tacoplot_stack_ly)
 export(tacoplot_venn)
