### 2024-07-15 #6313 @diazrenata
@kelly-sovacool and @david-barnett - just checking in to see how things are going and when you may have the chance to revisit `tidytacos`?
