### 2024-07-25 LebeerLab/tidytacos#28 @david-barnett
[BUG] read_tidytacos error handling broken
Hi tidytacos team, I'm testing out the package as part of the JOSS review and will raise issues for any bugs I spot on the way, such as this little one below.

cheers
David

typo in read_tidytacos code means cryptic error appears if count table not found, instead of what you intended

``` r
library(tidytacos)
urt %>% write_tidytacos("temp_dir")
unlink("temp_dir/counts.csv")
read_tidytacos("temp_dir")
#> Error in paste("File", counts, ", containing count data not found in", : cannot coerce type 'closure' to vector of type 'character'
```
