### 2024-07-25 LebeerLab/tidytacos#29 @david-barnett
[BUG] tacoplot_stack x argument doesn't accept string 
tacoplot_stack x argument requests a string in the help doc, but won't accept one

``` r
library(tidytacos)
urt %>% tacoplot_stack(x = "participant")
#> Warning in tacoplot_stack(., x = "participant"): Sample labels not unique,
#> samples are aggregated.
#> Error in `geom_bar()`:
#> ! Problem while computing aesthetics.
#> ℹ Error occurred in the 1st layer.
#> Caused by error in `forcats::fct_reorder()`:
#> ! length(f) == length(.x) is not TRUE

# As a side note: "Label not in sample table" seems like it should be an immediate error rather than a warning
urt %>% tacoplot_stack(x = participants)
#> Warning in tacoplot_stack(., x = participants): Label 'participants' not found
#> in the samples table.
#> Error in `pull()`:
#> Caused by error:
#> ! object 'participants' not found

```
