### 2024-07-25 LebeerLab/tidytacos#30 @david-barnett
tacoplot_stack pie = TRUE suggestions
To avoid misuse I feel the pie = TRUE should error if more than one sample is supplied. Or at least have an example of reasonable use.

``` r
library(tidytacos)
urt %>%
  filter_samples(sample_id == "s169") %>%
  tacoplot_stack(pie = TRUE)
```

![](https://i.imgur.com/3RsSkdd.png)

``` r

urt %>%
  filter_samples(participant == "CON83") %>%
  tacoplot_stack(pie = TRUE)
```

![](https://i.imgur.com/1QRpPB8.png)

``` r

urt %>%
  filter_samples(location == "N", method == "S") %>%
  tacoplot_stack(pie = TRUE)
```

![](https://i.imgur.com/9mNWpOI.png)
