### 2024-07-25 8b0ac67
Tacoplot stack handles strings or objects as label
