### 2024-07-25 0e165b7
Deal with function like names as label
For example sample
