### 2024-07-25 20d394c
Allow sample_clustered as argument for x
