### 2024-07-26 LebeerLab/tidytacos#32 @david-barnett
No examples for the filter/mutate/select functions
None of the functions in handlers.R have examples on their help pages.

https://github.com/LebeerLab/tidytacos/blob/master/R/handlers.R

You might like to combine a few of the related help pages together, e.g. all the filtering functions,

But I definitely think you need to give at least one working example for each function, or otherwise direct users who look in the help where to find examples elsewhere.
