### 2024-07-26 LebeerLab/tidytacos#33 @david-barnett
[BUG] tacoplot_ord x argument requests string, but does the wrong thing with a string
I'd suggest changing the docs instead of the code here, and to give an example of use.

``` r
library(tidytacos)

urt %>% tacoplot_ord(x = "participant")
#> Warning in tacoplot_ord(., x = "participant"): Empty samples detected, removing
#> them from the analysis
```

![](https://i.imgur.com/Z7EC04J.png)
