### 2024-07-26 LebeerLab/tidytacos#34 @david-barnett
[BUG] tacoplot_ord silently ignores colouring by sample or sample_id
If this is intended behaviour I would expect it to warn me it will not colour each separate sample.

``` r
library(tidytacos)

urt %>% tacoplot_ord(x = sample)
#> Warning in tacoplot_ord(., x = sample): Empty samples detected, removing them
#> from the analysis
```

![](https://i.imgur.com/sSV04hb.png)

``` r

urt %>% tacoplot_ord(x = sample_id)
#> Warning in tacoplot_ord(., x = sample_id): Empty samples detected, removing
#> them from the analysis
```

![](https://i.imgur.com/vdyCQIb.png)
