### 2024-07-26 LebeerLab/tidytacos#35 @david-barnett
several function arguments not documented - and rcmdcheck warnings ignored
`add_metadata` has a `metadata_tibble` argument but this is documented as `metadata`

This should give you a warning when you run rcmdcheck so i'm not sure how your CI is passing...

Edit: so i checked your github actions setup and see you've configured it to only fail on errors. You have a lot of undocumented argument warnings <https://github.com/LebeerLab/tidytacos/actions/runs/10097946221/job/27923934246#step:6:358>

I can understand not wanting to fail on notes, as your use of tidy evaluation gives a lot of rogue notes, but ignoring these warnings isn't a good idea, e.g. here you miss documentation problems.
