### 2024-07-26 LebeerLab/tidytacos#32 @david-barnett
this is actually a more general problem. perhaps most functions don't have examples on their help pages.

I feel you need to add examples on help pages, or direct to specific places in online documentation
