### 2024-07-26 LebeerLab/tidytacos#37 @david-barnett
[BUG] rank_names function unexpected behaviour
rank_names should "Return rank names associated with a tidytacos object"

But really it just always returns `c("kingdom", "phylum", "class", "order", "family", "genus")`

Which is quite unexpected.

``` r
library(tidytacos)

x <- matrix(c(1500, 1300, 280, 356), ncol = 2)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")
my_taco <- create_tidytacos(x, taxa_are_columns = FALSE)
my_taco$taxa
#> # A tibble: 2 × 2
#>   taxon  taxon_id
#>   <chr>  <chr>
#> 1 taxon1 t1
#> 2 taxon2 t2

rank_names(my_taco)
#> [1] "kingdom" "phylum"  "class"   "order"   "family"  "genus"
```

rank_names is called indirectly by tacoplot_stack, which then causes this kind of problem

``` r
my_taco %>% tacoplot_stack()
#> Error in `mutate()`:
#> ℹ In argument: `best_classification = purrr::pmap_chr(...)`.
#> Caused by error in `ta$taxa[, rank_names]`:
#> ! Can't subset columns that don't exist.
#> ✖ Columns `kingdom`, `phylum`, `class`, `order`, `family`, etc. don't exist.
```
