### 2024-07-26 LebeerLab/tidytacos#37 @david-barnett
Possibly one route towards an overall solution here could be to explicitly explain what ranks tidytacos expects a tidytacos object to have. You could check for this and throw sensible errors, either at the point of attempting to create a tidytacos class object that lacks the mandatory ranks, or when trying to use any function that would need them.
