### 2024-07-26 LebeerLab/tidytacos#37 @david-barnett
and also good to explain the special handling of the species rank somewhere obvious
