### 2024-07-26 LebeerLab/tidytacos#38 @david-barnett
tacoplot_alphas: samples with missing alpha diversity values removed only at plotting stage
tacoplot_ord notifies the user and drops empty samples up front, maybe tacoplot_alphas could do something similar?

When I applied tacoplot_alphas directly on a tidytacos object it was quite mysterious to have samples dropped from the plot.

Or maybe a warning in add_alpha is also a good idea, saying something like: N samples were empty, returning NA for those samples.

``` r
library(tidytacos)
urt_s <- urt %>%
  filter_samples(method == "S") %>%
  add_alphas(methods = "shannon")

urt_s %>% tacoplot_alphas(group_by = location)
#> Warning: Removed 2 rows containing non-finite outside the scale range
#> (`stat_ydensity()`).
#> Warning: Removed 2 rows containing missing values or values outside the scale range
#> (`geom_point()`).
```

![](https://i.imgur.com/QVBZztp.png)

``` r

sum(is.na(urt_s$samples$shannon))
#> [1] 2
```
