### 2024-07-26 #6313 @david-barnett
### Summary

I've completed a second round of review, which is the first time I assessed the software itself in depth. I'm pleased with the additional tutorial material available on the pkgdown website, and I still find the base design of tidytacos quite beginner-friendly. I worked through the quick start, stacked barplots, and data processing tutorials, which gave me a good walkthrough of the core functionality of the package. I encountered a few bugs when experimenting with the tools introduced in the tutorials and have filed GitHub issues for these.

My main criticism at this point concerns function-level documentation. Many functions are minimally described on their help pages with no examples of use. Additionally, many functions are not covered in the tutorials, making them difficult to use without reading the underlying code.

I would also like to see a more detailed description/justification of the required structure of the taxonomy table, such as which ranks must be included and the special treatment given to particular variables (e.g., species, taxon_id, sequence). This info would help prevent confusion when users try to import their own data into the tidytacos format.

Lastly, it would be beneficial to get R CMD check passing without warnings on the CI. There are currently documentation-related warnings, which are ignored.

### Minor Points on the Revised Article

- Line 37: The Ernst et al. reference seems malformed; it lacks a date.
- Line 59: "Each row represents such a read count" – I think you could more clearly explain the long format used.
- Line 71: The package is ggplot2.
- Line 72: "Additionally, also ..." is redundant.

The comparison with phyloseq in your getting started guide is quite compelling! Consider adding some of the points about the disadvantages of phyloseq's design to the paper and/or the homepage?
https://lebeerlab.github.io/tidytacos/articles/getting-started.html

### Other Notes

- Consider recommending remotes instead of devtools for github installation, as it's a lighter install for casual users.
