### 2024-07-29 12834b6
Separate infer rank_names to method
