### 2024-07-30 3bfe01f
Add test case for tacoplot stack string labels
