+++ tests/testthat/test-cases.R
@@ -3,6 +3,19 @@ test_that("Can plot ord after taxa aggregation", {
     # and the plot_ord function uses this to get the relative abundance
     urta <- urt %>% aggregate_taxa(rank="family") 
     expect_no_error(
-        urta %>% tacoplot_ord(x="plate")
+        expect_warning(
+            urta %>% tacoplot_ord(x="plate")
+        )
     )
+})
+
+test_that('Tacoplot stack accepts strings or objects', {
+
+    expect_no_error(
+        urt %>% tacoplot_stack(x="sample")
+    )
+    expect_no_error(
+        urt %>% tacoplot_stack(x=sample)
+    )
+
 })
