### 2024-07-30 fc85c49
Warn about removing empty samples and do so prior to add_alphas
