+++ tests/testthat/test-mantel.R
@@ -1,6 +1,6 @@
 test_that("Can perform Mantel test on one variable", {
 
-    expect_warning(
+    suppressWarnings(
         mantel <- urt %>% perform_mantel_test("plate")
     ) # empty samples
     expect_gt(mantel$signif, 0.1)
@@ -20,6 +20,10 @@ test_that("Can perform Mantel test on array of variables", {
 
 test_that("Throw error when trying to test one constant variable", {
 
-    expect_error(urt %>% perform_mantel_test("run"))
+    expect_error(
+        suppressWarnings(
+            urt %>% perform_mantel_test("run")
+        )
+    )
     
 })
