+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   getting-started: getting-started.html
   network-analysis: network-analysis.html
   stacked-barplot: stacked-barplot.html
-last_built: 2024-07-25T16:50Z
+last_built: 2024-07-30T14:29Z
 urls:
   reference: https://lebeerlab.github.io/tidytacos/reference
   article: https://lebeerlab.github.io/tidytacos/articles
