### 2024-07-30 531eabc
Add optional pkgs to suggests
