+++ DESCRIPTION
@@ -35,6 +35,16 @@ Suggests:
     phyloseq,
     plotly,
     MCL,
+    ggpubr,
+    jsonlite,
+    seqinr,
+    umap,
+    pheatmap,
+    rstatix,
+    Rtsne,
+    eulerr,
+    fastDummies,
+    SpiecEasi,
     Matrix,
     igraph,
     rmarkdown,
