+++ R/plotters.R
@@ -5,15 +5,10 @@
 # @param ta a tidytacos object
 # @param n an integer
 #
-prepare_for_bp <- function(ta, n = 12, extended = TRUE, ...) {
+prepare_for_bp <- function(ta, n = 12, extended = TRUE, order_by=NULL) {
 
-  extParams <- list()
-  if (length(list(...))) {
-    extParams <- unlist(list(...))
-  }
   # custom order
-  if (!is.null(extParams[["order_by"]])) {
-    order_by <- extParams[["order_by"]]
+  if (!is.null(order_by)) {
     ta$samples <- ta$samples %>% arrange(order_by)
     ta$samples$sample_clustered <- as.factor(ta$samples[[order_by]])
   }
@@ -73,10 +68,8 @@ get_ord_stat <- function(ta, x, stat.method, distance=distance) {
 #' Default is FALSE, as pie chart representations can be misleading to interpret.
 #' @param order_by an optional column name to order the samples by. 
 #' For examples order_by=sample would order the x-axis by the sample names instead of by similar profiles.
-#' @param ... Extra arguments to prepare the plot. Currently only `order_by` is supported.
-#' 
 #' @export
-tacoplot_stack <- function(ta, n = 12, x = sample_clustered, pie = FALSE, ...) {
+tacoplot_stack <- function(ta, n = 12, x = sample_clustered, pie = FALSE, order_by=NULL) {
   # convert promise to formula
 
   
@@ -113,7 +106,7 @@ tacoplot_stack <- function(ta, n = 12, x = sample_clustered, pie = FALSE, ...) {
   }
 
   # make plot and return
-  plot <- prepare_for_bp(ta, n, extended=T, ...) %>%
+  plot <- prepare_for_bp(ta, n, extended=T, order_by=order_by) %>%
     ggplot(aes(
       x = forcats::fct_reorder(!!x, as.integer(sample_clustered)),
       y = rel_abundance, fill = taxon_name_color)) +
@@ -153,8 +146,9 @@ tacoplot_stack <- function(ta, n = 12, x = sample_clustered, pie = FALSE, ...) {
 #' @param x A string, representing the column name used to label the x-axis
 #' @param order_by an optional column name to order the samples by. 
 #' For examples order_by=sample would order the x-axis by the sample names instead of by similar profiles.
+#' 
 #' @export
-tacoplot_stack_ly <- function(ta, n = 12, x = sample_clustered, ...) {
+tacoplot_stack_ly <- function(ta, n = 12, x = sample_clustered, order_by=NULL) {
   force_optional_dependency("plotly")
   # convert promise to formula
   x <- rlang::enquo(x)
@@ -163,7 +157,7 @@ tacoplot_stack_ly <- function(ta, n = 12, x = sample_clustered, ...) {
   plot <- rlang::eval_tidy(rlang::quo_squash(
     quo({
       # make plot and return
-      prepare_for_bp(ta, n, extended=T, if(length(list(...))) enquos(...) else NULL) %>%
+      prepare_for_bp(ta, n, extended=T, order_by=order_by) %>%
         plotly::plot_ly(
           x = ~forcats::fct_reorder(!!x, as.integer(sample_clustered)),
           y = ~rel_abundance,
@@ -202,6 +196,7 @@ tacoplot_stack_ly <- function(ta, n = 12, x = sample_clustered, ...) {
 #' @param palette A vector of colors, used as the palette for coloring sample
 #' @param title a string to display as title of the plot.
 #'   groups.
+#' @param ... Extra arguments to pass to the add_ord function.
 #'
 #' @export
 tacoplot_ord_ly <- function(ta, x=NULL, samplenames = sample_id, ord="pcoa", dims=2, 
[…]
