+++ R/varia.R
@@ -44,6 +44,7 @@ process_count_selection <- function(ta) {
 
 #' Return rank names associated with a tidytacos object if these are defined. In case of undefined rank names, the function returns the taxon_id field.
 #'
+#' @param ta a tidytacos object
 #' @export
 rank_names <- function(ta) {
 
