### 2024-07-30 5471218
Remove oldrel-2 for ubuntu
