+++ .github/workflows/R-CMD-check.yaml
@@ -32,8 +32,6 @@ jobs:
           # - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest,   r: 'release'}
           - {os: ubuntu-latest,   r: 'oldrel-1'}
-          - {os: ubuntu-latest,   r: 'oldrel-2'}
-
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
