### 2024-07-31 86041c4
Raise error when trying pcoa on <3 samples (due to k-1 max dimension representation)
