+++ man/filter_samples.Rd
@@ -11,6 +11,19 @@ filter_samples(ta, ...)
 
 \item{...}{Filter criteria for the samples table.}
 }
+\value{
+A tidytacos object.
+}
 \description{
 Filter the samples
+}
+\examples{
+
+# subset urt to keep only nasopharynx samples
+urt_nf <- urt \%>\% filter_samples(location == "NF")
+# subset urt to keep only samples from plate 1 and 2
+urt_plate_1_2 <- urt \%>\% filter_samples(plate \%in\% c(1,2))
+# subset the blanks in leaf
+leaf_blanks <- leaf \%>\% filter_samples(startsWith(description, "BLANK"))
+
 }
