+++ man/aggregate_samples.Rd
@@ -9,6 +9,9 @@ aggregate_samples(ta)
 \arguments{
 \item{ta}{A tidytacos object.}
 }
+\value{
+A tidytacos object.
+}
 \description{
 \code{aggregate_samples} merges sample content of samples which have identical values for all columns in the sample table (except sample_id).
 }
