+++ man/cluster_samples.Rd
@@ -11,6 +11,9 @@ cluster_samples(ta, n_clusters)
 
 \item{n_clusters}{The number of desired clusters.}
 }
+\value{
+A tidytacos object with a cluster column in the samples table.
+}
 \description{
 \code{cluster_samples} clusters the samples into n clusters and adds these
 clusters to a new variable "cluster" in the sample table.
