+++ man/add_prevalence.Rd
@@ -32,10 +32,10 @@ Supply condition as a string.
 \examples{
 # add prevalences of all taxa
 urtp <- urt \%>\% add_prevalence()
-urtp$taxa \%>\% select(taxon_id, prevalence)
+urtp$taxa \%>\% dplyr::select(taxon_id, prevalence)
 
 # add prevalences and fisher test for location
 urtpf <- urt \%>\% add_prevalence(condition="location", fisher_test=T, relative=T)
-urtpf$taxa \%>\% select(taxon_id, prevalence_in_N, prevalence_in_NF, fisher_p)
+urtpf$taxa \%>\% dplyr::select(taxon_id, prevalence_in_N, prevalence_in_NF, fisher_p)
 
 }
