+++ R/adders_taxa.R
@@ -334,7 +334,7 @@ add_jervis_bardy <- function(ta, dna_conc, sample_condition = T, min_pres = 3) {
 #' urtp$taxa %>% dplyr::select(taxon_id, prevalence)
 #' 
 #' # add prevalences and fisher test for location
-#' urtpf <- urt %>% add_prevalence(condition="location", fisher_test=T, relative=T)
+#' urtpf <- urt %>% add_prevalence(condition="location", fisher_test=TRUE, relative=TRUE)
 #' urtpf$taxa %>% dplyr::select(taxon_id, prevalence_in_N, prevalence_in_NF, fisher_p)
 #' 
 #' @export
