+++ man/add_prevalence.Rd
@@ -35,7 +35,7 @@ urtp <- urt \%>\% add_prevalence()
 urtp$taxa \%>\% dplyr::select(taxon_id, prevalence)
 
 # add prevalences and fisher test for location
-urtpf <- urt \%>\% add_prevalence(condition="location", fisher_test=T, relative=T)
+urtpf <- urt \%>\% add_prevalence(condition="location", fisher_test=TRUE, relative=TRUE)
 urtpf$taxa \%>\% dplyr::select(taxon_id, prevalence_in_N, prevalence_in_NF, fisher_p)
 
 }
