+++ R/adders_taxa.R
@@ -92,7 +92,7 @@ classify_taxa <- function(
 #' @examples 
 #' urt_g <- urt %>% add_taxon_name()
 #' # add the species name if present (which is often uncertain in amplicon data)
-#' urt_s <- urt %>% add_taxon_name(include_species = T)
+#' urt_s <- urt %>% add_taxon_name(include_species = TRUE)
 #' @importFrom stats na.omit
 #' @export
 add_taxon_name <- function(
