### 2024-08-08 LebeerLab/tidytacos#46 @kelly-sovacool
[BUG] minor warnings during R CMD build
**Describe the bug**

R CMD build emits warnings during installation:

```
── R CMD build ────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/private/var/folders/yb/f86k3qcj10nfyx57x30_dgnnsds0t8/T/RtmpTemN5J/remotes178f71dc179bf/LebeerLab-tidytacos-61b451d/DESCRIPTION’ ...
─  preparing ‘tidytacos’:
─  checking DESCRIPTION meta-information ...Warning in person1(given = given[[i]], family = family[[i]], middle = middle[[i]],  :
     It is recommended to use ‘given’ instead of ‘middle’.
    OK
   Warning in person1(given = given[[i]], family = family[[i]], middle = middle[[i]],  :
     It is recommended to use ‘given’ instead of ‘middle’.
─  building ‘tidytacos_1.0.0.tar.gz’
   Warning: invalid uid value replaced by that for user 'nobody'
   Warning: invalid gid value replaced by that for user 'nobody'
```

Installation still proceeds successfully, so this is not a blocking problem.

**To Reproduce**

in a fresh environment:

```R
devtools::install_github("LebeerLab/tidytacos")
```

or with a local clone of the repo:

```R
devtools::install("path/to/tidytacos")
```

**Expected behavior**

No warnings

**Screenshots**

**Version information (please complete the following information):**
 - OS: [e.g. iOS, ubuntu, windows] ```Sys.info()["sysname"]``` **macOS 14.5**
 - R version [e.g. 4.1.2] ```R.version``` **4.3.1**
 - tidytacos version [e.g. 0.2.2] ```packageVersion('tidytacos')``` **1.0.0 -- cloned from master branch**

**Additional context**
