### 2024-08-08 LebeerLab/tidytacos#47 @kelly-sovacool
[BUG] R CMD Check failing on local clone of master branch
**Describe the bug**

R CMD Check is failing on a local clone of the master branch.

```
══ Documenting ════════════════════════════════════════════════════════════════════════════════════════════
ℹ Updating tidytacos documentation
ℹ Loading tidytacos
Error in loadNamespace(x) : there is no package called ‘SpiecEasi’
```

**To Reproduce**

Clone the repo, run `devtools::install()` followed by `devtools::check()` in a fresh environment.

**Expected behavior**

Check completing with no errors, warnings, or notes.

**Screenshots**

**Version information (please complete the following information):**
 - OS: [e.g. iOS, ubuntu, windows] ```Sys.info()["sysname"]```
 - R version [e.g. 4.1.2] ```R.version```
 - tidytacos version [e.g. 0.2.2] ```packageVersion('tidytacos')```

**Additional context**

SpiecEasi is not in your DESCRIPTION file. I notice your gh actions workflow [installs SpiecEasi along with other packages manually](https://github.com/LebeerLab/tidytacos/blob/61b451dceb66f38bd708bfa9ca61b1bb1aa769a6/.github/workflows/R-CMD-check.yaml#L47-L61). Any of these packages that are actually used by the packages should be added to the DESCRIPTION as Imports or Suggests. I recommend removing them from the `extra-packages` part of your setup so you'll have a more accurate check workflow.

In case you need it, here's a guide on specifying how to install packages from github, bioconductor, etc in your DESCRIPTION file: <https://cran.r-project.org/web/packages/devtools/vignettes/dependencies.html>
