### 2024-08-08 LebeerLab/tidytacos#49 @kelly-sovacool
community guidelines missing from README
JOSS requirement:

> There should be clear guidelines for third-parties wishing to:
>
> - Contribute to the software
> - Report issues or problems with the software
> - Seek support

<https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines>
