### 2024-08-08 LebeerLab/tidytacos#50 @kelly-sovacool
recommend adding example data files for documentation
The quickstart explains how to load files into tidytacos:

```R
taco <- read_tidytacos("/path/to/my_data")
```

but then proceeds to use the `urt` dataset for the rest of the tutorial, which is already a tidytacos object.

It would be helpful to have an example file in your package in `inst/extdata` and use it in the tutorial, so users will have a better understanding of the expected file format.

Additionally, it would be best practice to have code in `data-raw` to show how `urt` and `leaf` were created. <https://r-pkgs.org/data.html#sec-data-data-raw>
