### 2024-08-08 LebeerLab/tidytacos#51 @kelly-sovacool
docs style: referring to function names
As a general rule, I recommend referring to specific functions as if they were proper names, followed by parentheses and enclosed in backticks. This helps make documentation more concise and easy to read.

For example, this sentence:

> The add_total_count function will add total sample read numbers to the sample table.

Could be revised as:

> `add_total_count()` will add total sample read numbers to the sample table.

As an added bonus, pkgdown will be able to turn that function reference into a hyperlink to its doc.

This is a soft recommendation.
