### 2024-08-08 LebeerLab/tidytacos#52 @kelly-sovacool
docs need instructions for users to import their own raw data
The quick start guide directs users to the source code to learn about more options for importing data.

> More options to import and convert your data can be found [here](https://github.com/LebeerLab/tidytacos/blob/dev/R/converters.R).

I strongly discourage this. Novice users should not be expected to read the source code to learn how to use the package. This information should be distilled into a vignette for users to read.

(_Originally posted by @kelly-sovacool in https://github.com/LebeerLab/tidytacos/issues/50#issuecomment-2275929329_)
