+++ data-raw/data-urt-rda.R
@@ -1,0 +1,3 @@
+library(tidytacos)
+
+urt <- read_tidytacos("tidytacos/urt")
