+++ data-raw/data-leaf-rda.R
@@ -1,0 +1,3 @@
+library(tidytacos)
+
+leaf <- read_tidytacos("tidytacos/leaf")
