### 2024-08-08 LebeerLab/tidytacos#53 @kelly-sovacool
missing unit tests for some functions
I recommend having at least some unit tests for every function, and especially so for user-facing functions. It is not necessary to have 100% code coverage, but having minimal tests for user-facing functions helps safe guard against accidentally introducing API-breaking changes.

From looking at the codecov reports, it appears several functions lack any unit tests. These include:

- from_dada
- from_phyloseq
- create_biom_header
- to_biom
- add_jervis_bardy
- get_ord_stat
- tacoplot_zoom
- tacoplot_venn
- tacoplot_venn_ly
- tacoplot_euler
- tacoplot_prevalences
