### 2024-08-08 #6313 @kelly-sovacool
# Review summary

tidytacos is a compelling new R package for microbial ecology and microbiome analysis.
The paper is well-written and makes a great case for the need the package fills in this space
and its usefulness for novices and advanced users alike.

I found the package was easy to install (although there were minor warnings),
and the vignettes are overall helpful for demonstrating the package's capabilities.
However, there is a critical lack of detail on how users can import their own
raw data into the package, as the getting started vignette links users to the
source code directly.
This is a significant barrier to entry for novice users who wish to analyze their own data.
Additionally, as noted by @david-barnett, the documentation lacks examples for many of the functions.

While conducting my review of the package I opened up several issues.
I separated these into lists of "Blocking issues" and "Suggestions" below.
I leave it up to the discretion of the authors to decide whether to address the
suggestions, but in my view the blocking issues do need to be addressed before publication.

Overall, this is a great package that will be a valuable addition to the field,
it just needs minor updates before it will be fully ready for publication.

## Blocking issues

- https://github.com/LebeerLab/tidytacos/issues/49
- https://github.com/LebeerLab/tidytacos/issues/52

## Suggestions

- https://github.com/LebeerLab/tidytacos/issues/46
- https://github.com/LebeerLab/tidytacos/issues/47
- https://github.com/LebeerLab/tidytacos/issues/50
- https://github.com/LebeerLab/tidytacos/issues/51
- https://github.com/LebeerLab/tidytacos/issues/53
