### 2024-08-14 ac22357
Add plot tests
