### 2024-08-14 0ca35be
Add NEWS.md
