+++ NEWS.md
@@ -1,0 +1,3 @@
+# tidytacos 1.0.0
+
+* Initial submission.
