+++ NEWS.md
@@ -1,3 +1,0 @@
-# tidytacos 1.0.0
-
-* Initial submission.
