+++ DESCRIPTION
@@ -6,7 +6,7 @@ Authors@R: c(
     person("Stijn", "Wittouck", , "stijn.wittouck@uantwerpen.be", role = c("aut", "cre")),
     person("Tim", "Van Rillaer", , "tim.vanrillaer@uantwerpen.be", role = c("aut")),
     person("Wenke", "Smets", , "wenke.smets@uantwerpen.be", role = c("aut")),
-    person("Sarah", "Lebeer", "sarah.lebeer@uantwerpen.be", role = c("fnd")))
+    person("Sarah", "Lebeer", , "sarah.lebeer@uantwerpen.be", role = c("fnd")))
 Description: A toolkit for processing and visualizing taxonomic composition data
     of microbial communities, based on the tidyverse framework.
 License: GPL-3 + file LICENSE
