### 2024-08-14 df5e4b6
Add getting started section
