+++ R/network.R
@@ -1,5 +1,5 @@
 #' Perform network inference with SparCC on a tidytacos object,
-#' after dropping rare taxa. See \link[SpiecEasi]{sparcc}.
+#' after dropping rare taxa. See [SpiecEasi::sparcc()].
 #'
 #' @param ta a tidytacos object
 #' @param min_occurrence Percentage of samples the taxon
@@ -67,15 +67,15 @@ network <- function(ta,
   network.out
 }
 
-#' Filters the output of \code{\link{network}} to a minimal threshold
+#' Filters the output of [network()] to a minimal threshold
 #' and transforms to matrix for downstream clustering or heatplot visualization.
 #'
 #'
-#' @param network network object generated by \code{\link{network}}.
+#' @param network network object generated by [network()].
 #' @param threshold absolute value of correlations below this threshold are
 #' @param fdr the threshold for false discovery rate
 #' (if pvalues are calculated for the network) filtered out.
-#' @param ... Extra arguments to pass to the \code{\link[stat]{p.adjust}} function.
+#' @param ... Extra arguments to pass to the [stat::p.adjust()] function.
 #' 
 #'@export
 filter_network <- function(network, threshold = 0.1, fdr = NULL, ...) {
@@ -108,7 +108,7 @@ filter_network <- function(network, threshold = 0.1, fdr = NULL, ...) {
   as.matrix(network_filtered)
 }
 
-#' Performs Markov Clustering on a sparcc network matrix generated by \code{\link{filter_network}}. 
+#' Performs Markov Clustering on a sparcc network matrix generated by [filter_network()]. 
 #' and returns a tibble with clusters and taxon identifiers. Optionally, the network can be visualized.
 #'
 #' @param network network network in matrix form.
@@ -119,6 +119,8 @@ filter_network <- function(network, threshold = 0.1, fdr = NULL, ...) {
 cluster_network <- function(network, min_n = 3, visualize = F) {
   force_optional_dependency("MCL")
   force_optional_dependency("Matrix")
+  taxa_per_clust <- NULL
+
   network[network < 0] <- 0
   res <- MCL::mcl(network, ESM = T, addLoops=T)
   Matrix::diag(res$Equilibrium.state.matrix) <- 0
@@ -197,7 +199,7 @@ pca_taxa <- function(ta, cluster_name, taxon_name=taxon, sample_name=sample){
 
 
 #' Calculates eigentaxa values based on SparCC - MCL generated clusters per sample. 
-#' It is advised to run \code{\link{cluster_taxa}} on the tidyamplicons object
+#' It is advised to run [cluster_taxa()] on the tidytacos object
 #' before running this function to add the clusters if you want to stray from any default parameters.
 #'
 #' @param ta a tidytacos object.
