+++ R/tidytacos-package.R
@@ -26,11 +26,12 @@
 #' @import stringr
 #' @import vegan
 #' @importFrom tibble tibble
+#' @importFrom stats dist
 globalVariables(c(
     "sample_id", "samples",
     "taxon_id", "taxa", "taxon", "taxon_name", "taxon_name_color",  
     "count", "rel_abundance", "presence", "alpha_metrics",
-    "sample_clustered", "prevalence", "cluster",
+    "sample_clustered", "prevalence", "cluster", "total_absolute_abundance",
     "spike_count", "total_count", "mean_rel_abundance", "spike_abundance",
     "total_count", "prevalence", "total_density",
     ".", "packageVersion", "packageDescription", "ord1", "ord2",
