+++ man/add_tree.Rd
@@ -1,0 +1,34 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/adders_tree.R
+\name{add_tree}
+\alias{add_tree}
+\title{Construct a phylogenetic tree from the sequences in the taxa table of a tidytacos object.}
+\usage{
+add_tree(
+  ta,
+  sequence_var = sequence,
+  outgroup_taxon_id = NULL,
+  aln_args = list(),
+  tree_fit_args = list()
+)
+}
+\arguments{
+\item{ta}{A tidytacos object.}
+
+\item{sequence_var}{The name of the column in the taxa table that contains the sequence}
+
+\item{aln_args}{An optional list of arguments to pass to the [DECIPHER::AlignSeqs()] function.}
+
+\item{tree_fit_args}{An optional list of arguments to pass to the [phangorn::optim.pml()] function.}
+
+\item{method}{The method by which to arrange the taxon names. Currently only
+mean_rel_abundance.}
+}
+\value{
+A tidytacos object with a phylotree slot.
+}
+\description{
+`add_tree()` performs a multiple sequence alignment on the ASV sequences in
+the taxa table and then constructs a phylogenetic tree. 
+The tree is added to the tidytacos object under a variable called `tree`.
+}
