+++ man/calculate_unifrac_distances.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/adders_samples.R
+\name{calculate_unifrac_distances}
+\alias{calculate_unifrac_distances}
+\title{Calculate unifrac distance matrix from a tidytacos object with a rooted tree}
+\usage{
+calculate_unifrac_distances(ta, ...)
+}
+\description{
+Calculate unifrac distance matrix from a tidytacos object with a rooted tree
+}
