+++ NAMESPACE
@@ -25,12 +25,14 @@ export(add_taxon_name_color)
 export(add_total_absolute_abundance)
 export(add_total_count)
 export(add_total_density)
+export(add_tree)
 export(aggregate_samples)
 export(aggregate_taxa)
 export(as_label)
 export(as_name)
 export(as_phyloseq)
 export(betas)
+export(calculate_unifrac_distances)
 export(classify_taxa)
 export(cluster_network)
 export(cluster_samples)
