### 2024-08-23 74c3dd1
Improve documentation add_tree and unifrac calc
