+++ man/calculate_unifrac_distances.Rd
@@ -6,6 +6,11 @@
 \usage{
 calculate_unifrac_distances(ta, ...)
 }
+\arguments{
+\item{ta}{A tidytacos object with a rooted tree in the "tree" slot.}
+
+\item{...}{Additional arguments to pass to the [phyloseq::UniFrac()] function.}
+}
 \description{
 Calculate unifrac distance matrix from a tidytacos object with a rooted tree
 }
