+++ NAMESPACE
@@ -97,6 +97,7 @@ import(stringr)
 import(tidyr)
 import(vegan)
 importFrom(magrittr,"%>%")
+importFrom(methods,as)
 importFrom(readr,read_csv)
 importFrom(readr,write_csv)
 importFrom(rlang,":=")
@@ -116,5 +117,6 @@ importFrom(stats,na.omit)
 importFrom(stats,p.adjust)
 importFrom(stats,prcomp)
 importFrom(stats,t.test)
+importFrom(stats,update)
 importFrom(stats,wilcox.test)
 importFrom(tibble,tibble)
