+++ README.md
@@ -2,7 +2,7 @@ tidytacos <img src="man/figures/logo.png" align="right" width="200"/>
 ======================
 [![R-CMD-check](https://github.com/LebeerLab/tidytacos/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/LebeerLab/tidytacos/actions/workflows/R-CMD-check.yaml)
 [![codecov](https://codecov.io/gh/LebeerLab/tidytacos/graph/badge.svg?token=532YS16DXU)](https://codecov.io/gh/LebeerLab/tidytacos)
-
+<br><br><br><br><br>
 ## Overview
 
 Tidytacos (tidy TAxonomic COmpositionS) is an R package for the exploration of microbial community data. Such community data consists of read counts generated by amplicon sequencing (e.g. a region of the 16S rRNA gene) or metagenome (shotgun) sequencing. Each read count represents a number of sequencing reads identified for some taxon (an ASV, OTU, species, or higher-level taxon) in a sample. 
