### 2024-08-30 LebeerLab/tidytacos#49 @wsmets
Thanks for pointing this out. README is updated.
