+++ man/add_alpha.Rd
@@ -21,13 +21,14 @@ tidytacos object.
 \details{
 This function can add different alpha diversity measures to the sample table, specified by the method argument.
 The following methods are available:
-- invsimpson: Inverse Simpson index
-- shannon: Shannon index
-- simpson: Simpson index
-- pielou: Pielou's evenness index
-- obs: Observed richness
-- s.chao1: Chao1 richness estimator
-- s.ace: ACE richness estimator
+
+* invsimpson: Inverse Simpson index
+* shannon: Shannon index
+* simpson: Simpson index
+* pielou: Pielou's evenness index
+* obs: Observed richness
+* s.chao1: Chao1 richness estimator
+* s.ace: ACE richness estimator
 }
 \examples{
 # Initiate counts matrix
