### 2024-08-30 ff7ec68
Finalize add_tree
