### 2024-08-30 cea793f
Add unifrac as possible distance metric for ord calculation (#55)
* Add tree generation and unifrac calculation

* Fix braces

* Improve documentation add_tree and unifrac calc

* Finalize add_tree

---------

Co-authored-by: Stijn <wittouck_stijn@hotmail.com>
