+++ NAMESPACE
@@ -25,12 +25,14 @@ export(add_taxon_name_color)
 export(add_total_absolute_abundance)
 export(add_total_count)
 export(add_total_density)
+export(add_tree)
 export(aggregate_samples)
 export(aggregate_taxa)
 export(as_label)
 export(as_name)
 export(as_phyloseq)
 export(betas)
+export(calculate_unifrac_distances)
 export(classify_taxa)
 export(cluster_network)
 export(cluster_samples)
@@ -95,6 +97,7 @@ import(stringr)
 import(tidyr)
 import(vegan)
 importFrom(magrittr,"%>%")
+importFrom(methods,as)
 importFrom(readr,read_csv)
 importFrom(readr,write_csv)
 importFrom(rlang,":=")
@@ -114,5 +117,6 @@ importFrom(stats,na.omit)
 importFrom(stats,p.adjust)
 importFrom(stats,prcomp)
 importFrom(stats,t.test)
+importFrom(stats,update)
 importFrom(stats,wilcox.test)
 importFrom(tibble,tibble)
