### 2024-08-30 edf0112
Minor points on the revised article
