+++ publication/joss/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-![](tacologo2.png)
+![](tacologo2.png){ width=120px }
 
 The `tidytacos` (tidy TAxonomic COmpositionS) package is an R package for the exploration of microbial community data. Such community data consists of read counts generated by amplicon sequencing (e.g. a region of the 16S rRNA gene) or metagenome (shotgun) sequencing. `tidytacos` is based on the tidy principle introduced by Hadley Wickham where the data are stored in a consistent format [@wickham2023r]. Specifically, `tidytacos` uses the tidy format and grammar to select, transform, and prepare microbial community data for visualization and analyses. In addition, it offers a range of functions for popular and less known analyses and visualizations of microbial community data. `tidytacos` is designed for researchers across various levels of expertise, to offer both an improved accessibility to microbial community data and to allow easy transformation of data for novel approaches of visualization and analysis.  
 
@@ -46,10 +46,10 @@ In the past two decades, DNA sequencing for the determination of microbial commu
 # Functionality
 
 `tidytacos` is a package for R, which is the most widely used programming language for microbial community analyses. It adheres to the tidy principles of coding and of structuring data and builds on the tidyverse collection of packages [@wickham2019welcome]. The package stores microbial community data in one “tidytacos object” which is a list of three tables: 
-1)	The counts: these are the counts of reads for each taxon (OTU/ASV/phylotype) in each sample. Each row represents such a read count. 
+1)	The counts: these are the counts of reads for each taxon (OTU/ASV/phylotype) in each sample. Each row contains exactly one read count, representing a single observation. This is the typical long format in the tidyverse. 
 2)	The samples: this table contains the sample metadata. Each row represents a sample. 
 3)	The taxa: this table contains the taxonomy and other metadata for the taxa. Each row represents a taxon. 
-The package is called `tidytacos` because each of the tables is tidy: each row represents an observation and each column a variable. Additionally, it adopts the grammar for data manipulation introduced by the `dplyr` package of the tidyverse, allowing, for example, the filtering and selection of samples and taxa and their associated information (e.g. filter_samples, filter_taxa, select_samples, select_taxa). Also, aggregating samples or taxa (aggregate_samples, aggregate_taxa) and conversion of the taxon counts (add_rel_abundance/add_clr_abundance) are supported, including a function to determine absolute abundances (add_absolute_abundance) based on the spike-in method [@smets2016method; @wang2021current]. There are quick functions for visualizing community composition of samples using stacked bar plots (tacoplot_stack) or visualizing a two-dimensional ordination of samples (tacoplot_ord) based on the ggplot package of the tidyverse. Additionally, also functions to study the effect of sample variables on community variation are available (e.g. perform_adonis) and differentially abundant taxa can be determined with a method based on the logratio differences introduced by @aitchison1982statistical. The full overview of functions, code, a quick start tutorial, and more advanced tutorials are available on the [Github](https://github.com/LebeerLab/tidytacos) page, https://github.com/LebeerLab/tidytacos.
[…]
