### 2024-09-02 9ca037c
Bugfix: return ta in infer_rank_names
