+++ tests/testthat/test-cases.R
@@ -20,13 +20,13 @@ test_that('Tacoplot stack accepts strings or objects', {
 
 })
 
-test_that("Aggregation works when correcting rank names", {
+test_that("Infer rank names bug (#57)", {
 
     t <- test_taco()
     t$taxa$domain <- c("Dom","Dom")
     t$taxa$phylum <- c("Phyl","Phyl")
     t$taxa$class <- c("Class1","Class2")
     t <- t %>% set_rank_names(c("domain","phylum","class"))
-    t <- t %>% aggregate_taxa(rank="class")
+    t <- t %>% aggregate_taxa(rank="phylum")
     expect_equal(nrow(t$taxa), 1)
 })
