+++ tests/testthat/test-cases.R
@@ -27,6 +27,6 @@ test_that("Aggregation works when correcting rank names", {
     t$taxa$phylum <- c("Phyl","Phyl")
     t$taxa$class <- c("Class1","Class2")
     t <- t %>% set_rank_names(c("domain","phylum","class"))
-    t <- t %>% aggregate_taxa(rank="class")
+    t <- t %>% aggregate_taxa(rank="phylum")
     expect_equal(nrow(t$taxa), 1)
 })
