### 2024-09-02 fd235ee
Increment version number to 1.0.2
