### 2024-09-02 c56664e
Stage for release 1.0.2 (#58)
* Add getting started section

* Add by='terms' to adonis due to updated default

* Add unifrac as possible distance metric for ord calculation (#55)
