+++ publication/joss/paper.bib
@@ -279,11 +279,12 @@
 } 
 
 @article{ernst1mia, 
-	  title={mia: Microbiome Analysis. 2023}, 
+	  title={mia: Microbiome Analysis}, 
 	    author={Ernst, F and Shetty, S and Borman, T and Lahti, L}, 
 	      journal={R Package Version}, 
 	        volume={1}, 
-		  number={0} 
+		  number={0}, 
+	  		year={2023}
 } 
 
 @article{aitchison1982statistical, 
