+++ R/adders_samples.R
@@ -1,6 +1,6 @@
 #' Add metadata to the tidytacos object
 #'
-#' \code{add_metadata} adds sample or taxon metadata to the sample or taxon
+#' `add_metadata()` adds sample or taxon metadata to the sample or taxon
 #' table, respectively, of a tidytacos object.
 #'
 #' @param ta A tidytacos object.
@@ -59,7 +59,7 @@ add_metadata <- function(ta, metadata, table_type = "sample") {
 
 #' Add total read count per sample
 #'
-#' \code{add_total_count} adds the total read count per sample to the sample
+#' `add_total_count()` adds the total read count per sample to the sample
 #' table of a tidytacos object under the variable name total_count.
 #'
 #' @param ta A tidytacos object.
@@ -104,21 +104,22 @@ add_total_count <- function(ta) {
 
 #' Add alpha diversity measure
 #'
-#' \code{add_alpha} adds an alpha diversity measures to the sample table of a
+#' `add_alpha()` adds an alpha diversity measures to the sample table of a
 #' tidytacos object.
 #'
 #' This function can add different alpha diversity measures to the sample table, specified by the method argument.
 #' The following methods are available:
-#' - invsimpson: Inverse Simpson index
-#' - shannon: Shannon index
-#' - simpson: Simpson index
-#' - pielou: Pielou's evenness index
-#' - obs: Observed richness
-#' - s.chao1: Chao1 richness estimator
-#' - s.ace: ACE richness estimator
+#'
+#' * invsimpson: Inverse Simpson index
+#' * shannon: Shannon index
+#' * simpson: Simpson index
+#' * pielou: Pielou's evenness index
+#' * obs: Observed richness
+#' * s.chao1: Chao1 richness estimator
+#' * s.ace: ACE richness estimator
 #'
 #' @param ta A tidytacos object.
-#' @param method The diversity measure to use, see \code{\link[vegan]{diversity}} for further information on these.
+#' @param method The diversity measure to use, see [vegan::diversity()] for further information on these.
 #' @return A tidytacos object with the alpha diversity measure added.
 #' @examples
 #' # Initiate counts matrix
@@ -144,6 +145,7 @@ add_total_count <- function(ta) {
 #' @export
 add_alpha <- function(ta, method="invsimpson") {
 
+  value <- NULL
   vegan_standard_methods <- c("invsimpson","shannon","simpson")
   vegan_estimateR_methods <- c("obs", "s.chao1", "s.ace")
 
@@ -191,12 +193,12 @@ add_alpha <- function(ta, method="invsimpson") {
 
 #' Add alpha diversity measures
 #'
-#' \code{add_alpha} adds selected alpha diversity measures to the sample table of a
+#' `add_alpha()` adds selected alpha diversity measures to the sample table of a
 #' tidytacos object.
 #'
 #' This function can add multiple different alpha diversity measures to the sample table, specified by the methods argument.
 #' @param ta A tidytacos object.
-#' @param methods A character vector of the diversity measure to use, see \code{\link[tidytacos]{add_alpha}} for examples.
+#' @param methods A character vector of the diversity measure to use, see [add_alpha()] for examples.
 #' Optionally use 'all' to add all diversity measures.
 #' @return A tidytacos object with the selected alpha diversity measures added.
 #' @examples
@@ -226,7 +228,7 @@ calculate_alpha_pielou <- function(ta) {
 
 #' Add clustering-based sample order
 #'
-#' \code{add_sample_clustered} adds a new variable defining a sample order based
[…]
