### 2024-09-03 fc76ecb
Internalize func doc
