### 2024-09-03 8b88227
Document internal fnc
