### 2024-09-03 #6313 @wsmets
Thanks again for diving into the package, @david-barnett and @kelly-sovacool and giving us constructive feedback. Also, we appreciate your appreciation. We have looked into all your points of advice. All GitHub issues were addressed and closed. In short we:
-     added instructions to the [README ](https://github.com/LebeerLab/tidytacos) for contributing to the software, reporting issues, and asking questions
-     explained how to import data from various sources in more detail on the [README ](https://github.com/LebeerLab/tidytacos) page
-     tackled the lack of function-level documentation by adding examples for most of the functions, including all functions in [handlers.R ](https://github.com/LebeerLab/tidytacos/blob/master/R/handlers.R)
-     decided that adding more tutorials remains a work in progress to cover all functionalities of the tidytacos package. For now, the existing tutorials and examples already cover most of the functionality.
-     now expect R CMD check to pass without warnings
-     addressed the minor points for the article text as advised
-     decided not to include large, specialized packages like SpiecEasi in standard installation as not everyone would use the functions associated with them and we would like to keep the installation brief for general use.

We hope this satisfies everyone and, if not, are happy to receive any feedback.
