+++ publication/joss/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-![](tacologo2.png){ width=120px }
+![](tacologo2.png){ width=220px }
 
 The `tidytacos` (tidy TAxonomic COmpositionS) package is an R package for the exploration of microbial community data. Such community data consists of read counts generated by amplicon sequencing (e.g. a region of the 16S rRNA gene) or metagenome (shotgun) sequencing. `tidytacos` is based on the tidy principle introduced by Hadley Wickham where the data are stored in a consistent format [@wickham2023r]. Specifically, `tidytacos` uses the tidy format and grammar to select, transform, and prepare microbial community data for visualization and analyses. In addition, it offers a range of functions for popular and less known analyses and visualizations of microbial community data. `tidytacos` is designed for researchers across various levels of expertise, to offer both an improved accessibility to microbial community data and to allow easy transformation of data for novel approaches of visualization and analysis.  
 
@@ -40,7 +40,7 @@ The `tidytacos` (tidy TAxonomic COmpositionS) package is an R package for the ex
 # Statement of need
 
 In the past two decades, DNA sequencing for the determination of microbial community composition has emerged as a powerful and transformative tool in the field of microbiology. Advancements in sequencing technologies have made it possible to explore and understand the vast diversity of microbial communities present in various environments (including the human microbiome) with unprecedented accuracy and depth at increasingly lower costs. This has led to more and more researchers, many with no prior experience in this field, attempting to determine the microbial community composition of samples using amplicon and metagenome sequencing technologies. Unsurprisingly, there is a high need for accessible tools both to process raw sequencing reads to taxonomically annotated read counts and subsequently interpret these taxonomic compositions. Researchers having to handle this kind of data for the first time often struggle with the gathering, transformation, and preparation of data for visualisation and analyses, which is a crucial step in the data analysis process. These steps lay the foundation of accurate and reliable insights.
-`tidytacos` is a toolkit for data wrangling of DNA-sequencing-based microbial community data. It increases accessibility of microbial community data, but also allows researchers who are more experienced with this type of data to address novel questions. The tool includes a range of simple functions for common exploratory analyses, visualizations, and statistical analyses. `tidytacos` offers similar functionalities as packages such as microViz (@barnett2021microviz), mia (@ernst1mia), and phyloseq (@mcmurdie2013phyloseq), but was designed specifically to be very accessible to researchers with limited R and general programming experience. To achieve this, tidytacos was embedded in the tidyverse framework, which provides a style of programming that is very intuitive for new (but also experienced) programmers and is often the starting point for many people learning R. Tidytacos is also very easy to install, since it has only the main tidyverse packages as dependencies for the vast majority of its functionality. Finally, tidytacos also serves the needs of researchers who are looking to implement some rarely used or new visualizations and methods of analysis. `tidytacos` has already been used in a number of scientific publications by different researchers [@smets2023leaf; @smets2022bacterial; @legein2022greenhouse; @de2017comparing; @oerlemans2020impact; @de2019anterior; @oerlemans2022impacts; @jorissen2021case; @lebeer2022selective; @de2022randomized; @allonsius2019microbiome; @de2020lactobacilli; @spacova2022spontaneous; @ahannach2021microbial; @de2021nasal; @papini2023boosting; @peng2020return; @van2021piloting; @van2021oxygen; @lebeer2023citizen] and has also been used by several undergraduate, graduate, and PhD students during their research projects [@van2022resource; @legein2022phyllosphere], demonstrating its versatility in topics and user experience.   
[…]
