### 2024-09-04 13ec7da
Align logo to right
