+++ R/adders_samples.R
@@ -14,16 +14,16 @@
 #' @examples
 #' # Initiate counts matrix
 #' x <- matrix(
-#'  c(1500, 1300, 280, 356),
-#'  ncol = 2
+#'   c(1500, 1300, 280, 356),
+#'   ncol = 2
 #' )
 #' rownames(x) <- c("taxon1", "taxon2")
 #' colnames(x) <- c("sample1", "sample2")
 #'
 #' # Convert to tidytacos object
 #' data <- create_tidytacos(x,
-#'                      taxa_are_columns = FALSE
-#'                      )
+#'   taxa_are_columns = FALSE
+#' )
 #'
 #' # Initiate sample tibble
 #' sample <- c("sample1", "sample2")
@@ -32,20 +32,19 @@
 #'
 #' # Add sample tibble to tidytacos object
 #' data <- data %>%
-#' add_metadata(sample_tibble)
+#'   add_metadata(sample_tibble)
 #'
 #' # Initiate taxon tibble
-#' genera <- c("Lactobacillus","Limosilactobacillus")
-#' species <- c("crispatus","reuteri")
-#' taxonomy <- tibble::tibble(taxon=rownames(x),genera,species)
-#' 
+#' genera <- c("Lactobacillus", "Limosilactobacillus")
+#' species <- c("crispatus", "reuteri")
+#' taxonomy <- tibble::tibble(taxon = rownames(x), genera, species)
+#'
 #' # Add taxon tibble to tidytacos object
 #' data <- data %>%
-#'   add_metadata(taxonomy, table_type="taxa") 
+#'   add_metadata(taxonomy, table_type = "taxa")
 #'
 #' @export
 add_metadata <- function(ta, metadata, table_type = "sample") {
-
   if (table_type == "sample") {
     purrr::modify_at(ta, "samples", left_join, metadata)
   } else if (table_type == "taxa") {
@@ -54,7 +53,6 @@ add_metadata <- function(ta, metadata, table_type = "sample") {
   } else {
     stop("table_type must be either 'sample' or 'taxa'")
   }
-
 }
 
 #' Add total read count per sample
@@ -67,24 +65,23 @@ add_metadata <- function(ta, metadata, table_type = "sample") {
 #' @examples
 #' # Initiate counts matrix
 #' x <- matrix(
-#'  c(1500, 1300, 280, 356),
-#'  ncol = 2
+#'   c(1500, 1300, 280, 356),
+#'   ncol = 2
 #' )
 #' rownames(x) <- c("taxon1", "taxon2")
 #' colnames(x) <- c("sample1", "sample2")
 #'
 #' # Convert to tidytacos object
 #' data <- create_tidytacos(x,
-#'                      taxa_are_columns = FALSE
-#'                      )
+#'   taxa_are_columns = FALSE
+#' )
 #'
 #' # Add total counts
 #' data <- data %>%
-#'  add_total_count()
+#'   add_total_count()
 #'
 #' @export
 add_total_count <- function(ta) {
-
   # make table with sample and total count
   lib_sizes <- ta$counts %>%
     group_by(sample_id) %>%
@@ -99,7 +96,6 @@ add_total_count <- function(ta) {
 
   # return ta object
   ta
-
 }
 
 #' Add alpha diversity measure
@@ -124,29 +120,28 @@ add_total_count <- function(ta) {
 #' @examples
 #' # Initiate counts matrix
 #' x <- matrix(
-#'  c(1500, 1300, 280, 356),
-#'  ncol = 2
+#'   c(1500, 1300, 280, 356),
+#'   ncol = 2
 #' )
 #' rownames(x) <- c("taxon1", "taxon2")
[…]
