+++ R/handlers.R
@@ -7,20 +7,20 @@
 #' @param n Subsample size for rarefying the community.
 #' @param replace Whether to replace the read after it has been selected for the subsample so it can be sampled again. Default is FALSE.
 #' @return A tidytacos object.
-#' 
+#'
 #' @examples
 #' # discard samples with less than 1000 reads
-#' urt_1000 <- urt %>% 
-#'   add_total_count() %>% 
+#' urt_1000 <- urt %>%
+#'   add_total_count() %>%
 #'   filter_samples(total_count >= 1000)
-#' 
+#'
 #' # then rarefy to 1000 reads
 #' urt_1000 <- urt_1000 %>% rarefy(1000)
-#' 
+#'
 #' @export
 rarefy <- function(ta, n, replace = F) {
   ta$counts <- try(
-      ta$counts %>%
+    ta$counts %>%
       group_by(sample_id) %>%
       mutate(
         count =
@@ -32,10 +32,13 @@ rarefy <- function(ta, n, replace = F) {
       ungroup()
   )
 
-  if (class(ta$counts)[[1]] == "try-error"){
+  if (class(ta$counts)[[1]] == "try-error") {
     stop(
-      paste("Rarefying failed. Make sure that all samples contain at least the minimum number of reads.\n",
-    "Or use replace = TRUE, to allow sampling with replacement."))
+      paste(
+        "Rarefying failed. Make sure that all samples contain at least the minimum number of reads.\n",
+        "Or use replace = TRUE, to allow sampling with replacement."
+      )
+    )
   }
 
   ta %>%
@@ -113,7 +116,6 @@ change_id_taxa <- function(ta, taxon_id_new) {
 #' @return A tidytacos object.
 #' @export
 aggregate_samples <- function(ta) {
-
   sample_id_new <- NULL
 
   # sample table with only old and new sample names
@@ -157,17 +159,16 @@ aggregate_samples <- function(ta) {
 #' @param ta A tidytacos object.
 #' @param rank An optional rank to aggregate on.
 #' @return A tidytacos object.
-#' 
-#' @examples 
+#'
+#' @examples
 #' urt %>% aggregate_taxa(rank = "class")
 #' @export
 aggregate_taxa <- function(ta, rank = NULL) {
-
   taxon_id_new <- NULL
-  
+
   # Temporarily replace any NA's with strings as they interfere with aggregation
   ta$taxa[is.na(ta$taxa)] <- "unknown"
-  
+
   if (!is.null(rank)) {
     rank_names <-
       rank_names(ta) %>%
@@ -229,11 +230,11 @@ aggregate_taxa <- function(ta, rank = NULL) {
     rank_names(ta) %>% intersect(names(ta$taxa))
   )
   # Add new unique taxon label
-  if (!is.null(rank)){
-    include_species = eval(rank=="species")
-    ta <- ta %>% 
-      add_taxon_name(include_species = include_species) %>% 
-      mutate_taxa(taxon = taxon_name) %>% 
+  if (!is.null(rank)) {
+    include_species <- eval(rank == "species")
+    ta <- ta %>%
+      add_taxon_name(include_species = include_species) %>%
+      mutate_taxa(taxon = taxon_name) %>%
       suppressWarnings(select_taxa(-taxon_name))
   }
   ta
@@ -241,7 +242,7 @@ aggregate_taxa <- function(ta, rank = NULL) {
 
 #' Trim all sequences
 #'
-#' `trim_asvs()` trims sequence ends of the sequence supplied in the taxa table. 
+#' `trim_asvs()` trims sequence ends of the sequence supplied in the taxa table.
[…]
