+++ man/tacosum.Rd
@@ -16,8 +16,8 @@ tidytacos object.
 \examples{
 # Initiate counts matrix
 x <- matrix(
- c(1500, 1300, 280, 356),
- ncol = 2
+  c(1500, 1300, 280, 356),
+  ncol = 2
 )
 rownames(x) <- c("taxon1", "taxon2")
 colnames(x) <- c("sample1", "sample2")
