+++ man/betas.Rd
@@ -30,8 +30,8 @@ estimate for each combination of samples.
 \examples{
 # Initiate counts matrix
 x <- matrix(
- c(1500, 1300, 280, 356),
- ncol = 2
+  c(1500, 1300, 280, 356),
+  ncol = 2
 )
 rownames(x) <- c("taxon1", "taxon2")
 colnames(x) <- c("sample1", "sample2")
