+++ man/add_total_count.Rd
@@ -19,19 +19,19 @@ table of a tidytacos object under the variable name total_count.
 \examples{
 # Initiate counts matrix
 x <- matrix(
- c(1500, 1300, 280, 356),
- ncol = 2
+  c(1500, 1300, 280, 356),
+  ncol = 2
 )
 rownames(x) <- c("taxon1", "taxon2")
 colnames(x) <- c("sample1", "sample2")
 
 # Convert to tidytacos object
 data <- create_tidytacos(x,
-                     taxa_are_columns = FALSE
-                     )
+  taxa_are_columns = FALSE
+)
 
 # Add total counts
 data <- data \%>\%
- add_total_count()
+  add_total_count()
 
 }
