### 2024-09-04 49aed6a
Remove alignment logo
As it doesn't seem to work in pdf form anyway
