### 2024-09-04 462b675
Change logo size
