### 2024-09-04 b557c45
Resizing logo
