### 2024-10-22 LebeerLab/tidytacos#60 @david-barnett
create_tidytacos example shows incorrect (?) warning `Not all default rank names found`
https://lebeerlab.github.io/tidytacos/reference/create_tidytacos.html

create_tidytacos example shows a warning `Not all default rank names found` - but this warning cannot be right, because all default rank names **are** given in the example taxonomy table, yes?

I think the check in this line should be the other way around?

https://github.com/LebeerLab/tidytacos/blob/b557c45d2d03d971c23072fa0e56eada11de2491/R/converters.R#L178
