### 2024-10-22 LebeerLab/tidytacos#61 @david-barnett
Expecting multiple warnings in tests to fix check warnings?
I noticed that devtools test and your CI checks still produce warnings - https://github.com/LebeerLab/tidytacos/actions/runs/10700922653/job/29665720454#step:6:241

You can expect multiple warnings by nesting expect_warnings calls - e.g. see example below modifying test-mantel.R to fix the warning linked above

I would also recommend using the regexp argument, to ensure you notice behaviour changes in the future (throwing other unexpected warnings).

```r
test_that("Can perform Mantel test on array of variables", {
  skip_if_not_installed("fastDummies")
  expect_warning(
    expect_warning(
      expect_warning(
        mantel <- urt %>% perform_mantel_test(c("location", "method")),
        regexp = "Empty samples found"
      ),
      regexp = "Removed 3 empty samples"
    ),
    regexp = "non-square matrix"
  ) # empty samples

  expect_gt(mantel$signif, 0.1)
})
```
