### 2024-10-22 #6313 @david-barnett
Hello all,

I've completed my 3rd round of review:

1. I took a look through my previous issues, and I can confirm that most appear satisfactorily resolved - with a couple of relatively minor exceptions (see below)
2. I was now able to tick off every item on the [📝 checklist](https://github.com/openjournals/joss-reviews/issues/6313#issuecomment-1942273942)

I think that my currently outstanding issues are now relatively minor, and should not block publication, but I would still recommend the authors take a look.

## My previous issues that I can confirm are resolved:
- ✅ https://github.com/LebeerLab/tidytacos/issues/28
- ✅ https://github.com/LebeerLab/tidytacos/issues/29
- ✅ https://github.com/LebeerLab/tidytacos/issues/30 - multi-sample pie now triggers a desirable error
- ✅ https://github.com/LebeerLab/tidytacos/issues/32 - examples added to most functions
- ✅ https://github.com/LebeerLab/tidytacos/issues/33 - confirmed docs changed to not request string
- ✅ https://github.com/LebeerLab/tidytacos/issues/34 - confirmed warning about not colouring samples added
- ✅ https://github.com/LebeerLab/tidytacos/issues/38 - add_alphas now removes empty samples

## My previous issues that were not fully resolved:

- ⚙️ https://github.com/LebeerLab/tidytacos/issues/35 - add_metadata arg named is changed, but the CI still has small issues with unhandled warnings -> see suggestion at https://github.com/LebeerLab/tidytacos/issues/61

- ⚠️ https://github.com/LebeerLab/tidytacos/issues/36 - unexpected warning now thrown in new create_tidytacos example -> https://github.com/LebeerLab/tidytacos/issues/60

- ❓https://github.com/LebeerLab/tidytacos/issues/37 - did you document the special treatment of the "species" rank anywhere? - you mentioned it as a TODO in pull request 39, but I didn't spot a resolution? https://github.com/LebeerLab/tidytacos/pull/39#issue-2435362816

----
Link to my previous review summary: <https://github.com/openjournals/joss-reviews/issues/6313#issuecomment-2252987538>
