+++ R/adders_samples.R
@@ -42,7 +42,8 @@
 #' # Add taxon tibble to tidytacos object
 #' data <- data %>%
 #'   add_metadata(taxonomy, table_type = "taxa")
-#'
+#' @return A tidytacos object with metadata columns added to the taxa or sample table.
+#' @family sample-modifiers
 #' @export
 add_metadata <- function(ta, metadata, table_type = "sample") {
   if (table_type == "sample") {
@@ -79,7 +80,8 @@ add_metadata <- function(ta, metadata, table_type = "sample") {
 #' # Add total counts
 #' data <- data %>%
 #'   add_total_count()
-#'
+#' @return A tidytacos object with the total read count per sample added.
+#' @family sample-modifiers 
 #' @export
 add_total_count <- function(ta) {
   # make table with sample and total count
@@ -103,7 +105,8 @@ add_total_count <- function(ta) {
 #' `add_alpha()` adds an alpha diversity measures to the sample table of a
 #' tidytacos object.
 #'
-#' This function can add different alpha diversity measures to the sample table, specified by the method argument.
+#' This function can add different alpha diversity measures
+#' to the sample table, specified by the method argument.
 #' The following methods are available:
 #'
 #' * invsimpson: Inverse Simpson index
@@ -115,7 +118,8 @@ add_total_count <- function(ta) {
 #' * s.ace: ACE richness estimator
 #'
 #' @param ta A tidytacos object.
-#' @param method The diversity measure to use, see [vegan::diversity()] for further information on these.
+#' @param method The diversity measure to use,
+#' see [vegan::diversity()] for further information on these.
 #' @return A tidytacos object with the alpha diversity measure added.
 #' @examples
 #' # Initiate counts matrix
@@ -137,7 +141,8 @@ add_total_count <- function(ta) {
 #'
 #' data <- data %>%
 #'   add_alpha(method = "shannon")
-#'
+#' @family sample-modifiers
+#' @family diversity-metrics
 #' @export
 add_alpha <- function(ta, method = "invsimpson") {
   value <- NULL
@@ -197,6 +202,8 @@ add_alpha <- function(ta, method = "invsimpson") {
 #' @return A tidytacos object with the selected alpha diversity measures added.
 #' @examples
 #' urt_all_alphas <- urt %>% add_alphas()
+#' @family sample-modifiers
+#' @family diversity-metrics
 #' @export
 add_alphas <- function(ta, methods = "all") {
   if (any(methods == "all")) {
@@ -235,6 +242,9 @@ calculate_alpha_pielou <- function(ta) {
 #' @examples
 #' urtc <- urt %>% add_sample_clustered()
 #' urtc$samples %>% dplyr::select(sample_id, sample_clustered)
+#' @return A tidytacos object with a new variable
+#' `sample_clustered` added to the sample table.
+#' @family sample-modifiers
 #' @export
 add_sample_clustered <- function(ta) {
   # if only one sample => no clustering
@@ -326,7 +336,10 @@ perform_umap <- function(ta, dist_matrix, dims = 2, ...) {
 
 #' Calculate unifrac distance matrix from a tidytacos object with a rooted tree
 #' @param ta A tidytacos object with a rooted tree in the "tree" slot.
-#' @param ... Additional arguments to pass to the [phyloseq::UniFrac()] function.
+#' @inheritDotParams phyloseq::UniFrac
+#' @return A distance matrix.
+#' @family distance-metrics
+#' @family unifrac-distance-functions
 #' @export
 calculate_unifrac_distances <- function(ta, ...) {
   ensure_tree(ta, rooted = TRUE)
[…]
