+++ man/perform_anosim.Rd
@@ -35,6 +35,9 @@ perform_anosim(ta, group, ...)
     package.}
   }}
 }
+\value{
+an object of class "anosim"
+}
 \description{
 \code{perform_anosim()} performs the anosim test for statistical difference
 between groups of samples. The null hypothesis is that there is no difference
@@ -45,5 +48,4 @@ between microbial communities in the groups of samples.
 perform_anosim(urt, method, dist = "jaccard")
 # no statistical difference based on the method column
 # (high significance value and R close to 0).
-
 }

[…]
