+++ tests/testthat/test-mantel.R
@@ -9,10 +9,18 @@ test_that("Can perform Mantel test on one variable", {
 
 test_that("Can perform Mantel test on array of variables", {
 
-    skip_if_not_installed("fastDummies")    
+    skip_if_not_installed("fastDummies")
     expect_warning(
-        mantel <- urt %>% perform_mantel_test(c("location","method"))
-    ) # empty samples
+        expect_warning(
+            expect_warning(
+                mantel <- urt %>% 
+                  perform_mantel_test(c("location","method")),
+                regexp = "Empty samples found"
+            ),
+        regexp = "Removed 3 empty samples"
+        ),
+        regexp = "non-square matrix"
+    )
     expect_gt(mantel$signif, 0.1)
 
 })
