### 2024-10-22 cd14ce2
Mention taxa defaults in documentation
