### 2024-10-22 5b585c2
Minor changes to tests and mentioning default tax ranks in documentation (#62)
* Documentation updates

* Expect warnings with regular expressions

* Mention taxa defaults in documentation
