+++ R/converters.R
@@ -2,11 +2,14 @@
 #'
 #' `create_tidytacos()` returns a tidytacos object given a numeric matrix.
 #'
-#' This function initiates a tidytacos object based on a numeric matrix. It
-#' will automatically create a dummy taxa table and sample table which will need to be
-#' updated using the function [add_metadata()].
-#' When the taxa table is updated, the rank names can be set using the function [set_rank_names()]
-#' to make the tidytacos object aware of the taxonomy levels.
+#' This function initiates a tidytacos object based on a numeric matrix.
+#' It will automatically create a dummy taxa table and sample table which
+#' will need to be updated using the function [add_metadata()].
+#' When the taxa table is updated, the rank names can be set
+#' using the function [set_rank_names()] to make the tidytacos object
+#' aware of the taxonomy level order (from high to low).
+#' The taxa table should contain at the very least one rank name.
+#' The default rank names used by tidytacos are "domain", "phylum", "class", "order", "family", "genus" and "species".
 #'
 #' @param counts_matrix Numerical matrix containing the count data.
 #' @param taxa_are_columns A logical scalar. Are the taxa defined in columns?
@@ -39,13 +42,16 @@
 #' data <- add_metadata(data, taxonomy, table_type = "taxa")
 #' # rank names are inferred from the taxa table, but can be set manually
 #' # if we're not happy with the inferred rank names.
-#' data <- set_rank_names(data, c("domain", "phylum", "class", "order", "family", "genus"))
+#' data <- set_rank_names(
+#'   data, c("domain", "phylum", "class", "order", "family", "genus")
+#' )
 #' @family import-methods
 #' @return A tidytacos object.
 #' @export
 create_tidytacos <- function(counts_matrix, taxa_are_columns = TRUE) {
+
   if (
-    !is.matrix(counts_matrix) |
+    !is.matrix(counts_matrix) ||
       !is.numeric(counts_matrix)
   ) {
     stop("first argument should be a numeric matrix")
@@ -53,10 +59,10 @@ create_tidytacos <- function(counts_matrix, taxa_are_columns = TRUE) {
 
   if (!taxa_are_columns) counts_matrix <- t(counts_matrix)
   if (counts_matrix %>% rownames() %>% length() == 0) {
-    rownames(counts_matrix) <- str_c("s", seq(1:dim(counts_matrix)[1]))
+    rownames(counts_matrix) <- str_c("s", seq_len(dim(counts_matrix)[1]))
   }
   if (counts_matrix %>% colnames() %>% length() == 0) {
-    colnames(counts_matrix) <- str_c("s", seq(1:dim(counts_matrix)[2]))
+    colnames(counts_matrix) <- str_c("s", seq_len(dim(counts_matrix)[2]))
   }
 
   counts_matrix <-
@@ -91,21 +97,19 @@ create_tidytacos <- function(counts_matrix, taxa_are_columns = TRUE) {
     tibble(taxon = .) %>%
     mutate(taxon_id = !!taxon_ids)
 
-  # set rank names
-  expected_rank_names <- colnames(ta$taxa)[!colnames(ta$taxa) %in% c("taxon", "taxon_id", "sequence")]
-
-
-
   ta
 }
 
 #' Write community data in tidytacos format
 #'
-#' `write_tidytacos()` saves the tidytacos object into 3 .csv files. This format allows easy loading of the tidytacos object using the [read_tidytacos()] function.
+#' `write_tidytacos()` saves the tidytacos object into 3 .csv files.
+#' This format allows easy loading of the tidytacos object using
+#' the [read_tidytacos()] function.
 #'
 #' @importFrom readr write_csv
[…]
