### 2024-10-22 #6313 @TheOafidian
Hi David, many thanks for the in depth reviews as always!

I've tackled your two raised issues and have made some more references to the expected taxa ranks for using tidytacos in the intended way.
