### 2024-10-23 #6313 @diazrenata
@kelly-sovacool I believe this is ready for you to revisit when you have the time!
