### 2024-10-23 LebeerLab/tidytacos#52 @kelly-sovacool
@TheOafidian as far as I can tell this is unresolved. Maybe I'm not looking at the most updated version? Can you point to a commit or version where this was fixed?
