### 2024-10-23 LebeerLab/tidytacos#50 @kelly-sovacool
I see you added raw data to `data-raw`. The R package convention is to place raw data files in `inst/extdata/` and include R scripts in `data-raw/` showing how datasets in `data/` were generated. `data-raw/` is not installed with the package, so users will not be able to access it unless it is in `inst/extdata/`.

Here's the R packages book chapter on raw data, which goes in `inst/extdata/`: https://r-pkgs.org/data.html#sec-data-extdata

Here's the R packages book chapter on including data in your package: https://r-pkgs.org/data.html#sec-data-sysdata
