### 2024-10-23 LebeerLab/tidytacos#52 @kelly-sovacool
I see you added [example pseudocode](https://github.com/LebeerLab/tidytacos/blob/5b585c2c140ee08474729d2dc52bec8ea9c88cc3/README.md?plain=1#L50-L55) to the README and an [improved description](https://github.com/LebeerLab/tidytacos/blob/5b585c2c140ee08474729d2dc52bec8ea9c88cc3/vignettes/getting-started.Rmd#L88-L94) of the expected data format for raw data in the quick start vignette.

These changes are an improvement, but in my view are not quite sufficient to solve this issue. I would recommend showing a working, real example of importing raw data using the csv files you include in the package.

The README could have something like this:

(note: the file paths below assume you will have fixed the data organization as described here https://github.com/LebeerLab/tidytacos/issues/50#issuecomment-2433581017)

```R
OTUtable <- readr::read_csv(system.file('extdata','urt','counts.csv'))
taxonomy <- readr::read_csv(system.file('extdata','urt','taxa.csv'))
sampledata <- readr::read_csv(system.file('extdata','urt','samples.csv'))

taco <- create_tidytacos(OTUtable, taxa_are_columns = FALSE)
taco <- taco%>%
  add_metadata(taxonomy, table_type="taxa")%>%
  add_metadata(sampledata, table_type="sample")
```

and the quick start vignette could have a highly similar example around [here](https://github.com/LebeerLab/tidytacos/blob/5b585c2c140ee08474729d2dc52bec8ea9c88cc3/vignettes/getting-started.Rmd#L98-L106) instead of the non-working file paths currently shown.

The key detail is showing working code with real example files so users can learn how to use the package by example.

Finally, I still strongly discourage linking to the source code in the quick start vignette (or in any user-facing docs) as [here](https://github.com/LebeerLab/tidytacos/blob/5b585c2c140ee08474729d2dc52bec8ea9c88cc3/vignettes/getting-started.Rmd#L86). You could opt instead to link to a page on the docs website, e.g. the section on [file handling](https://lebeerlab.github.io/tidytacos/reference/index.html#file-handling) in the reference.
