### 2024-10-23 LebeerLab/tidytacos#50 @kelly-sovacool
on a related note, this snippet in the README does not work for end users:

https://github.com/LebeerLab/tidytacos/blob/5b585c2c140ee08474729d2dc52bec8ea9c88cc3/README.md?plain=1#L60

once you move the example files to their proper locations as described above, you will be able to do something like this instead:

```R
taco <- read_tidytacos(system.file("extdata", "tidytacos", "leaf", package = "tidytacos"))
```
