### 2024-10-23 #6313 @kelly-sovacool
Hi all,

A couple of my issues have not been resolved. I added additional comments in the respective issue threads to explain more in depth what I am looking for. In my view, these issues should be straightforward to resolve and would help tidytacos have robust documentation to meet the JOSS requirement for example usage. Thus, I would highly encourage the authors to address them before publishing the package.

- https://github.com/LebeerLab/tidytacos/issues/52
- https://github.com/LebeerLab/tidytacos/issues/50
