+++ data-raw/data-urt-rda.R
@@ -1,3 +1,0 @@
-library(tidytacos)
-
-urt <- read_tidytacos("tidytacos/urt")
