### 2024-10-29 0027857
Add unittest for from_dada
