+++ R/converters.R
@@ -338,6 +338,10 @@ from_phyloseq <- function(ps) {
 #' @param taxa Taxa table, output of [dada2::assignTaxonomy()].
 #' @param taxa_are_columns A logical scalar. Are the taxa defined in columns?
 #' @return A tidytacos object.
+#' @examples
+#' seqtab <- system.file("extdata", "dada2", "seqtab.rds", package = "tidytacos")
+#' taxa <- system.file("extdata", "dada2", "taxa.rds", package = "tidytacos")
+#' taco <- from_dada(seqtab, taxa)
 #' @export
 from_dada <- function(seqtab, taxa, taxa_are_columns = TRUE) {
   if ("matrix" %in% class(seqtab)) {
