### 2024-10-30 8ae8506
Dev (#64)
* Documentation updates

* Expect warnings with regular expressions

* Mention taxa defaults in documentation

* Increment version number to 1.0.2.9000

* Increment version

* Add autolinks

* Add dada2 example data

* Add unittest for from_dada

* Read rds in example

* Add phyloseq example data

* Update getting-started.Rmd

* Update README.md

* Update getting-started.Rmd

* Document and rm linebreak
