+++ data-raw/data-urt-biom.R
@@ -1,4 +1,0 @@
-library(tidytacos)
-
-urt <- read_tidytacos("tidytacos/urt")
-urt %>% to_biom("biom/urt.biom")
