+++ data-raw/data-leaf-rda.R
@@ -1,3 +1,0 @@
-library(tidytacos)
-
-leaf <- read_tidytacos("tidytacos/leaf")
