### 2024-10-31 7a3a1f9
Update README.md in seqtab taxa are columns
