### 2024-11-05 3b3c03a
Dev (#66)
* Update README.md with readRDS

* Update README.md taxa in right format

* Update README.md added blank line

* Update README.md more efficent taxa conversion

* Rename sequence column

* Update README.md in seqtab taxa are columns

---------

Co-authored-by: TheOaphidian <tim.van.rillaer@hotmail.com>
