+++ .github/workflows/R-CMD-check.yaml
@@ -57,7 +57,6 @@ jobs:
               any::fastDummies
               any::ggVennDiagram
               bioc::dada2
-              bioc::phyloseq
               github::zdk123/SpiecEasi
           needs: check
 
