+++ .github/workflows/pkgdown.yaml
@@ -42,7 +42,6 @@ jobs:
             any::fastDummies
             any::ggVennDiagram
             bioc::dada2
-            bioc::phyloseq
             github::zdk123/SpiecEasi
           needs: website
 
