### 2024-11-27 788e445
Add inheritDotParams for betas
