### 2024-12-11 bc84b73
Add option to keep empty samples
