### 2024-12-17 83637a6
Add docs for new argument to add_alpha
