### 2024-12-17 e61e773
Add option to keep empty samples in tacoplot_alphas
