### 2025-01-03 d75c431
Add pseudocount to clr and default to rclr on absence of pseudocount
