+++ R/plotters.R
@@ -551,11 +551,14 @@ tacoplot_alphas <- function(ta, group_by, compare_means = FALSE, keep_empty_samp
 
 #' Return a heatmap of prevalence of taxa in groups of samples
 #'
-#' Return a heatmap of all taxa above a certain threshold prevalence per condition, clusters them and compares prevalences with a fisher_test.
+#' Return a heatmap of all taxa above a certain threshold prevalence
+#' per condition, clusters them and compares prevalences with a fisher_test.
 #' @examples
 #' urt %>%
 #'   aggregate_taxa(rank = "order") %>%
-#'   tacoplot_prevalences(location, cutoff = .1, treeheight_row = 0, cutree_rows = 4, fontsize = 6, cellwidth = 15)
+#'   tacoplot_prevalences(location, cutoff = .1,
+#'   treeheight_row = 0, cutree_rows = 4,
+#'   fontsize = 6, cellwidth = 15)
 #'
 #' @param ta A tidytacos object.
 #' @param condition The row name of the condition which rrevalences are to be compared.
