+++ NAMESPACE
@@ -64,6 +64,7 @@ export(rank_names)
 export(rarefy)
 export(read_tidytacos)
 export(rel_abundance_matrix)
+export(remove_duplicate_samples)
 export(remove_empty_samples)
 export(reset_ids)
 export(samples)
