### 2025-01-07 3dbc541
Add basic LDA modeling wrappers
