+++ NAMESPACE
@@ -32,6 +32,7 @@ export(as_label)
 export(as_name)
 export(as_phyloseq)
 export(betas)
+export(calculate_lda_perplexities)
 export(calculate_unifrac_distances)
 export(classify_taxa)
 export(cluster_network)
@@ -58,6 +59,7 @@ export(mutate_taxa)
 export(network)
 export(perform_adonis)
 export(perform_anosim)
+export(perform_lda)
 export(perform_mantel_test)
 export(prevalences)
 export(rank_names)
