+++ man/ldaplot_alignment.Rd
@@ -1,0 +1,49 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/lda_models.R
+\name{ldaplot_alignment}
+\alias{ldaplot_alignment}
+\title{Plot LDA topic alignment results}
+\usage{
+ldaplot_alignment(lda_alignment, show_product = TRUE, ...)
+}
+\arguments{
+\item{lda_alignment}{The output of \code{align_lda_topics()}.}
+
+\item{show_product}{Whether to show the product or transport alignment.}
+
+\item{...}{
+  Arguments passed on to \code{\link[alto:plot_alignment]{alto::plot_alignment}}
+  \describe{
+    \item{\code{x}}{(required) An alignment class object resulting from
+\code{align_topics}.}
+    \item{\code{rect_gap}}{(optional) A float describing how much vertical space to put
+between topics within the same model. The units correspond to topic masses.
+Defaults to 0.2.}
+    \item{\code{color_by}}{(optional) What should the color of topics and weights
+encode? Defaults to 'path'. Other possible arguments are 'coherence',
+'refinement', or 'topic'.}
+    \item{\code{model_name_repair_fun}}{(optional) How should names be repaired before plotting?}
+    \item{\code{label_topics}}{(optional, default = \code{FALSE}) A \code{logical}
+specifying if topics should be labeled with the \code{"color_by"} information.}
+    \item{\code{add_leaves}}{(optional, default = \code{FALSE}) A \code{logical}
+specifying if the topic composition of leave-topics should be printed.}
+    \item{\code{leaves_text_size}}{(optional, default = \code{10}) specifies the font
+size of leaves annotations in \code{pt} if \code{add_leaves} is \code{TRUE}.}
+    \item{\code{n_features_in_leaves}}{(optional, default = 3) specifies the maximum
+number of features that should be included in the leaves annotations
+if \code{add_leaves} is \code{TRUE}.}
+    \item{\code{min_feature_prop}}{(optional, default = 0.1) specifies the minimum
+proportion of a feature in a topic for that feature to be included in
+the leaves annotations if \code{add_leaves} is \code{TRUE}.}
+    \item{\code{top_n_edges}}{(optional, \code{integer}, default = \code{NULL}) specifies
+the number of edges that should be drawn between the topics of subsequent models.
+The \code{top_n_edges} with the highest weights are drawn. If \code{NULL}
+(default), all edges are drawn.}
+  }}
+}
+\value{
+A ggplot object.
+}
+\description{
+\code{ldaplot_alignment()} plots the alignment of LDA topics across models.
+}
