+++ NAMESPACE
@@ -28,6 +28,7 @@ export(add_total_density)
 export(add_tree)
 export(aggregate_samples)
 export(aggregate_taxa)
+export(align_lda_topics)
 export(as_label)
 export(as_name)
 export(as_phyloseq)
@@ -51,6 +52,8 @@ export(filter_samples)
 export(filter_taxa)
 export(from_dada)
 export(from_phyloseq)
+export(ldaplot_alignment)
+export(ldaplot_beta)
 export(mean_rel_abundances)
 export(merge_tidytacos)
 export(mutate_counts)
