### 2025-01-20 #6313 @kthyng
@wsmets I'm sorry this has gone unaddressed for so long! I will take over.

@kelly-sovacool It looks like you will be satisfied with the review at this point, given your [last comment](https://github.com/openjournals/joss-reviews/issues/6313#issuecomment-2433672339), but if you see this it'd be great if you can check off your last box (assuming you are in fact satisfied)!
