### 2025-01-20 #6313 @kthyng
@wsmets Please add these DOIs to your bib file, if they are correct.
