### 2025-01-21 bd0a888
Increment version number to 1.0.6
