+++ DESCRIPTION
@@ -18,6 +18,7 @@ BugReports:
 Depends: 
     R (>= 4.1.0)
 Imports:
+    alto,
     dplyr,
     forcats,
     ggplot2,
@@ -29,6 +30,7 @@ Imports:
     stringr,
     tibble,
     tidyr,
+    topicmodels,
     vegan
 Suggests:
     ape,
