### 2025-01-21 322eb30
Move dependency to suggests
