+++ DESCRIPTION
@@ -18,7 +18,6 @@ BugReports:
 Depends: 
     R (>= 4.1.0)
 Imports:
-    alto,
     dplyr,
     forcats,
     ggplot2,
@@ -30,10 +29,10 @@ Imports:
     stringr,
     tibble,
     tidyr,
-    topicmodels,
     vegan
 Suggests:
     ape,
+    alto,
     biomformat,
     Biostrings,
     compositions,
@@ -62,6 +61,7 @@ Suggests:
     SpiecEasi,
     testthat (>= 3.1.9),
     tidyverse,
+    topicmodels,
     umap,
     vdiffr
 VignetteBuilder: 
