### 2025-01-21 19b3b03
Add optional deps to workflow
