+++ .github/workflows/test-coverage.yaml
@@ -32,8 +32,10 @@ jobs:
                 bioc::dada2
                 any::MCL
                 any::Matrix
+                any::topicmodels
                 any::ggVennDiagram
                 github::zdk123/SpiecEasi
+                github::lasy/alto
           needs: coverage
 
       - name: Test coverage
