### 2025-01-21 3d47af1
Update to 1.0.6 (#70)
* Add option to keep empty samples

* Add docs for new argument to add_alpha

* Add option to keep empty samples in tacoplot_alphas

* Add pseudocount to clr and default to rclr on absence of pseudocount

* Throw error on duplicate sample_ids

* Clear previous results on rerun

* Add basic LDA modeling wrappers

* Add alto wrappers

* Update paper details

* Increment version number to 1.0.6

* Add dependency packages

* Move dependency to suggests

* Add optional deps to workflow

* Update test-coverage.yaml
