+++ publication/joss/paper.bib
@@ -12,7 +12,8 @@
   volume={4},
   number={43},
   pages={1686},
-  year={2019}
+  year={2019},
+  doi = {10.21105/joss.01686}
 }
 
 @article{smets2016method,
@@ -22,7 +23,10 @@
   volume={96},
   pages={145--151},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier}, 
+  ISSN={0038-0717}, 
+  url={http://dx.doi.org/10.1016/j.soilbio.2016.02.003}, 
+  DOI={10.1016/j.soilbio.2016.02.003}
 }
 
 @article{wang2021current,
@@ -33,16 +37,22 @@
   number={9},
   pages={1797},
   year={2021},
-  publisher={MDPI}
+  publisher={MDPI}, 
+  ISSN={2076-2607}, 
+  url={http://dx.doi.org/10.3390/microorganisms9091797}, 
+  DOI={10.3390/microorganisms9091797}
 }
 
 @article{smets2023leaf,
   title={Leaf side determines the relative importance of dispersal versus host filtering in the phyllosphere microbiome},
   author={Smets, Wenke and Chock, Mason K and Walsh, Corinne M and Vanderburgh, Caihong Qiu and Kau, Ethan and Lindow, Steven E and Fierer, Noah and Koskella, Britt},
-  journal={Mbio},
+  journal={mBio},
   pages={e01111--23},
   year={2023},
-  publisher={Am Soc Microbiol}
+  publisher={American Society for Microbiology},
+  ISSN={2150-7511}, 
+  url={http://dx.doi.org/10.1128/mbio.01111-23},
+  DOI={10.1128/mbio.01111-23}
 }
 
 @article{smets2022bacterial,
@@ -53,7 +63,10 @@
   number={2},
   pages={e02420--21},
   year={2022},
-  publisher={Am Soc Microbiol}
+  publisher={Am Soc Microbiol}, 
+  ISSN={2165-0497}, 
+  url={http://dx.doi.org/10.1128/spectrum.02420-21}, 
+  DOI={10.1128/spectrum.02420-21}
 }
 
 @article{legein2022greenhouse,
@@ -64,7 +77,10 @@
   number={4},
   pages={e01755--22},
   year={2022},
-  publisher={Am Soc Microbiol}
+  publisher={Am Soc Microbiol}, 
+  ISSN={2165-0497}, 
+  url={http://dx.doi.org/10.1128/spectrum.01755-22}, 
+  DOI={10.1128/spectrum.01755-22}
 }
 
 @article{de2017comparing,
@@ -74,7 +90,10 @@
   volume={8},
   pages={2372},
   year={2017},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA}, 
+  ISSN={1664-302X}, 
+  url={http://dx.doi.org/10.3389/fmicb.2017.02372}, 
+  DOI={10.3389/fmicb.2017.02372}
 }
 
 
@@ -86,7 +105,10 @@
   number={1},
   pages={7976},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  ISSN={2045-2322}, 
+  url={http://dx.doi.org/10.1038/s41598-020-64705-x}, 
+  DOI={10.1038/s41598-020-64705-x}
 }
 
 @article{de2019anterior,
@@ -97,7 +119,10 @@
   number={6},
   pages={10--1128},
   year={2019},
-  publisher={Am Soc Microbiol}
+  publisher={Am Soc Microbiol},
+  ISSN={2379-5042},
+  url={http://dx.doi.org/10.1128/msphere.00532-19},
[…]
