+++ man/remove_duplicate_samples.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{remove_duplicate_samples}
+\alias{remove_duplicate_samples}
+\title{Removes duplicate samples from the tidytacos object}
+\usage{
+remove_duplicate_samples(ta)
+}
+\arguments{
+\item{ta}{a tidytacos object}
+}
+\value{
+the tidytacos object minus the duplicate samples
+}
+\description{
+Will remove rows with the exact same metadata as another row but a different sample_id.
+}
