+++ .github/workflows/pkgdown.yaml
@@ -34,7 +34,6 @@ jobs:
             any::tidyverse
             any::plotly
             any::vdiffr
-            any::compositions
             any::MCL
             any::Matrix
             any::umap

[…]
