### 2025-01-22 #6313 @wsmets
Dear @kthyng, thank you for taking over!
-	We have checked the DOI's and added them to the references wherever possible (just not of the PhD thesises).
-	We have checked authors and affiliations: they are correct
-	The latest release version of tidytacos is v1.0.6
-	This release version is published on Zenodo: https://zenodo.org/records/14717030, with DOI 10.5281/zenodo.14717030
-	To the manuscript, we have added 4 more references that used our tool. Because these publications happened recently, they weren’t in there when we first submitted. Lines 46-52.

Let us know if there is anything else we can do.
