### 2025-01-23 1b44480
Fix reindexing bug
