### 2025-01-28 #6313 @kthyng
@wsmets I see in your Zenodo archive that instead of listing the co-authors in the same order and without designation, as in the JOSS paper, you have labeled them with roles. Is this an important item for you? It also brings up for the question of one being listed as a "sponsor" given our guidelines about authorship: https://joss.readthedocs.io/en/latest/submitting.html#authorship. Can you comment on this?
