### 2025-01-28 #6313 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. I see a few "r"s and I'm not sure if any of the Latin names need to be capitalized. Please check in detail.
