### 2025-01-30 #6313 @TheOafidian
> [@wsmets](https://github.com/wsmets) I see in your Zenodo archive that instead of listing the co-authors in the same order and without designation, as in the JOSS paper, you have labeled them with roles. Is this an important item for you? It also brings up for the question of one being listed as a "sponsor" given our guidelines about authorship: https://joss.readthedocs.io/en/latest/submitting.html#authorship. Can you comment on this?

Hi Kristen, I couldn't find how to add any co-authors without assigning roles on Zenodo so I just picked whatever I had found. The "sponsor" is actually just our PI which I wasn't sure what role to denote that with, so I've now adapted that to the "Supervisor" role.
I've also changed the order of authors to that on the paper. We're not really too invested in the "roles", so no need for reflecting this in the JOSS paper.
