### 2025-01-30 #6313 @kthyng
@TheOafidian In this case, you can remove those roles and instead just like more co-authors, in the same order and name as in the paper. This link might help: https://help.zenodo.org/docs/deposit/describe-records/creators/
