+++ publication/joss/paper.bib
@@ -1,10 +1,19 @@
+@article{allonsius2019microbiome,
+  title={The microbiome of the {I}nvertebrate model host \emph{Galleria mellonella} is <i>dominated</i> by \textit{Enterococcus}},
+  author={Allonsius, Camille Nina and Van Beeck, Wannes and De Boeck, Ilke and Wittouck, Stijn and Lebeer, Sarah},
+  journal={Animal Microbiome},
+  volume={1},
+  pages={1--7},
+  year={2019},
+  publisher={Springer},
+  DOI={10.1186/s42523-019-0010-6}
+}
 @book{wickham2023r,
   title={R for data science},
   author={Wickham, Hadley and {\c{C}}etinkaya-Rundel, Mine and Grolemund, Garrett},
   year={2023},
   publisher={" O'Reilly Media, Inc."}
 }
-
 @article{wickham2019welcome,
   title={Welcome to the Tidyverse},
   author={Wickham, Hadley and Averick, Mara and Bryan, Jennifer and Chang, Winston and McGowan, Lucy D'Agostino and Fran{\c{c}}ois, Romain and Grolemund, Garrett and Hayes, Alex and Henry, Lionel and Hester, Jim and others},
@@ -15,7 +24,6 @@
   year={2019},
   doi = {10.21105/joss.01686}
 }
-
 @article{smets2016method,
   title={A method for simultaneous measurement of soil bacterial abundances and community composition via 16S rRNA gene sequencing},
   author={Smets, Wenke and Leff, Jonathan W and Bradford, Mark A and McCulley, Rebecca L and Lebeer, Sarah and Fierer, Noah},
@@ -28,7 +36,6 @@
   url={http://dx.doi.org/10.1016/j.soilbio.2016.02.003}, 
   DOI={10.1016/j.soilbio.2016.02.003}
 }
-
 @article{wang2021current,
   title={Current applications of absolute bacterial quantification in microbiome studies and decision-making regarding different biological questions},
   author={Wang, Xiaofan and Howe, Samantha and Deng, Feilong and Zhao, Jiangchao},
@@ -42,7 +49,6 @@
   url={http://dx.doi.org/10.3390/microorganisms9091797}, 
   DOI={10.3390/microorganisms9091797}
 }
-
 @article{smets2023leaf,
   title={Leaf side determines the relative importance of dispersal versus host filtering in the phyllosphere microbiome},
   author={Smets, Wenke and Chock, Mason K and Walsh, Corinne M and Vanderburgh, Caihong Qiu and Kau, Ethan and Lindow, Steven E and Fierer, Noah and Koskella, Britt},
@@ -54,7 +60,6 @@
   url={http://dx.doi.org/10.1128/mbio.01111-23},
   DOI={10.1128/mbio.01111-23}
 }
-
 @article{smets2022bacterial,
   title={Bacterial succession and community dynamics of the emerging leaf phyllosphere in spring},
   author={Smets, Wenke and Spada, Lucia Maria and Gandolfi, Isabella and Wuyts, Karen and Legein, Marie and Muyshondt, Babette and Samson, Roeland and Franzetti, Andrea and Lebeer, Sarah},
@@ -68,7 +73,6 @@
   url={http://dx.doi.org/10.1128/spectrum.02420-21}, 
   DOI={10.1128/spectrum.02420-21}
 }
-
 @article{legein2022greenhouse,
   title={The Greenhouse Phyllosphere Microbiome and Associations with Introduced Bumblebees and Predatory Mites},
   author={Legein, Marie and Smets, Wenke and Wuyts, Karen and Bosmans, Lien and Samson, Roeland and Lebeer, Sarah},
@@ -82,7 +86,6 @@
[…]
