+++ publication/joss/paper.bib
@@ -1,5 +1,5 @@
 @article{allonsius2019microbiome,
-  title={The microbiome of the {I}nvertebrate model host \emph{Galleria mellonella} is <i>dominated</i> by \textit{Enterococcus}},
+  title={The microbiome of the invertebrate model host \textit{{G}alleria mellonella} is dominated by \textit{{E}nterococcus}},
   author={Allonsius, Camille Nina and Van Beeck, Wannes and De Boeck, Ilke and Wittouck, Stijn and Lebeer, Sarah},
   journal={Animal Microbiome},
   volume={1},
@@ -186,7 +186,7 @@
   DOI={10.2139/ssrn.3422481}
 }
 @article{spacova2022spontaneous,
-  title={Spontaneous riboflavin-overproducing \emph{Limosilactobacillus reuteri} for biofortification of fermented foods},
+  title={Spontaneous riboflavin-overproducing \textit{{L}imosilactobacillus reuteri} for biofortification of fermented foods},
   author={Spacova, Irina and Ahannach, Sarah and Breynaert, Annelies and Erreygers, Isabel and Wittouck, Stijn and Bron, Peter A and Van Beeck, Wannes and Eilers, Tom and Alloul, Abbas and Blansaer, Na{\"\i}m and others},
   journal={Frontiers in Nutrition},
   volume={9},
@@ -206,7 +206,7 @@
   DOI={10.1016/j.isci.2021.103306}
 }
 @article{de2021nasal,
-  title={The nasal mutualist Dolosigranulum pigrum AMBR11 supports homeostasis via multiple mechanisms},
+  title={The nasal mutualist \textit{{D}olosigranulum pigrum} AMBR11 supports homeostasis via multiple mechanisms},
   author={De Boeck, Ilke and Wittouck, Stijn and Martens, Katleen and Spacova, Irina and Cauwenberghs, Eline and Allonsius, Camille Nina and J{\"o}rissen, Jennifer and Wuyts, Sander and Van Beeck, Wannes and Dillen, Jelle and others},
   journal={Iscience},
   volume={24},
@@ -226,7 +226,7 @@
   DOI={10.1016/j.biortech.2022.128285}
 }
 @article{peng2020return,
-  title={Return-sludge treatment with endogenous free nitrous acid limits nitrate production and N2O emission for mainstream partial nitritation/Anammox},
+  title={Return-sludge treatment with endogenous free nitrous acid limits nitrate production and N_{2}O emission for mainstream partial nitritation/Anammox},
   author={Peng, Lai and Xie, Yankai and Van Beeck, Wannes and Zhu, Weiqiang and Van Tendeloo, Michiel and Tytgat, Tom and Lebeer, Sarah and Vlaeminck, Siegfried E},
   journal={Environmental Science \& Technology},
   volume={54},
@@ -279,7 +279,7 @@
   school={University of Antwerp}
 }
 @article{barnett2021microviz, 
-  title={microViz: an R package for microbiome data visualization and statistics}, 
+  title={microViz: an {R} package for microbiome data visualization and statistics}, 
   author={Barnett, David JM and Arts, Ilja CW and Penders, John}, 
   journal={Journal of Open Source Software}, 
   volume={6}, 
@@ -289,7 +289,7 @@
   DOI={10.21105/joss.03201}
 } 
 @article{mcmurdie2013phyloseq, 
-      title={phyloseq: an R package for reproducible interactive analysis and graphics of microbiome census data}, 
+      title={phyloseq: an {R} package for reproducible interactive analysis and graphics of microbiome census data}, 
       author={McMurdie, Paul J and Holmes, Susan}, 
       journal={PloS one}, 
       volume={8}, 
@@ -339,7 +339,7 @@
[…]
