### 2025-01-31 #6313 @TheOafidian
> [@TheOafidian](https://github.com/TheOafidian) In this case, you can remove those roles and instead just like more co-authors, in the same order and name as in the paper. This link might help: https://help.zenodo.org/docs/deposit/describe-records/creators/

Done, thanks for the help Kristen!
